### 2023-11-20 #618 @hanecakr
fellingdateR: Estimate, report and combine felling dates of historical tree-ring series
Submitting Author Name: Kristof Haneca
Submitting Author Github Handle: @hanecakr
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/hanecakr/fellingdateR
Version submitted:
Submission type: Standard

Version accepted: TBD

---

```
Package: fellingdateR
Title: Estimate, report and combine felling dates of historical tree-ring
    series
Version: 0.0.0.9003
    person("Kristof", "Haneca", , "Kristof.Haneca@vlaanderen.be",role = c("aut", "cre"),
    person("Koen", "Van Daele", role = "ctb",
    person("Ronald", "Visser", role = "ctb",
    )
Description: `fellingdateR` is an R package that aims to facilitate the
  analysis and interpretation of tree-ring data from wooden
  cultural heritage objects and structures. The package standarizes the process
  of computing and combining felling date estimates, both for individual and
  groups of related tree-ring series.
    ggtext,
    MASS,
    plyr,
    dplR
    covr,

```

The package aims to facilitate and standardize the computation of felling dates from dated tree-ring series. It covers all steps of processing measured ring-width series from raw data to the reporting of felling date estimates, for single series and for groups of related tree-ring series.

Maily professional tree-ring scientists (dendrochronologists) that work on historical timbers and wooden objects (archaeology, architectural history, sculptures, panel paintings, etc.)

No, no other R-packages do the same job.

yes

All checks pass

- [x ] I have read the [rOpenSci packaging guide](https://devguide.ropensci.org/building.html).
- [x ] I have read the [author guide](https://devdevguide.netlify.app/authors-guide.html) and I expect to maintain this package for at least 2 years or to find a replacement.
