### 2023-11-21 #618 @mpadge
@hanecakr The error was because your URL field was markdown-style: "\[url\](url)", and should just be plain "url". I've edited to fix; you should now be able to call `@ropensci-review-bot check package` yourself to restart checks. Sorry for any inconvencience.
