+++ tests/testthat/test-sw_sum.R
@@ -1,45 +1,31 @@
 testthat::test_that("Output is a data.frame", {
-     x <- sw_sum(fellingdateR:::dummy7, plot = FALSE)
-     testthat::expect_s3_class(x, "data.frame")
-     testthat::expect_true('SPD' %in% names(x))
+        x <- sw_sum(dummy7, plot = FALSE)
+        testthat::expect_s3_class(x, "data.frame")
+        testthat::expect_true('SPD' %in% names(x))
 })
 
 testthat::test_that("Output is a data.frame", {
-     x <- sw_sum(fellingdateR:::dummy6, plot = FALSE)
-     testthat::expect_s3_class(x, "data.frame")
-     testthat::expect_true('SPD' %in% names(x))
+        x <- sw_sum(dummy6, plot = FALSE)
+        testthat::expect_s3_class(x, "data.frame")
+        testthat::expect_true('SPD' %in% names(x))
 })
 
 testthat::test_that("sw_sum does not work with invalid density function", {
-     testthat::expect_error(
-          sw_sum(
-               fellingdateR:::dummy7,
-               densfun = "nuka-cola"
-          ),
-          regexp = "not a supported distribution"
-     )
+        testthat::expect_error(sw_sum(dummy7,
+                                      densfun = "nuka-cola"),
+                               regexp = "not a supported distribution")
 })
 
 testthat::test_that("sw_sum does not work with invalid credMass", {
-     testthat::expect_error(
-          sw_sum(
-               fellingdateR:::dummy7,
-               credMass = "ninety percent"
-          )
-     )
+        testthat::expect_error(sw_sum(dummy7,
+                                      credMass = "ninety percent"))
 })
 
 
 testthat::test_that("sw_sum warning series without waney edge removed", {
-     testthat::expect_warning(
-          sw_sum(
-               fellingdateR:::dummy4)
-     )
+        testthat::expect_warning(sw_sum(dummy4))
 })
 
 testthat::test_that("sw_sum warning no series wtih sapwood or waney edge removed", {
-     testthat::expect_error(
-          suppressWarnings(sw_sum(
-               fellingdateR:::dummy5))
-     )
+        testthat::expect_error(suppressWarnings(sw_sum(dummy5)))
 })
