+++ tests/testthat/test-sw_combine_plot.R
@@ -1,5 +1,5 @@
 # Test if the function returns a ggplot object
 testthat::test_that("sw_combine_plot() returns a ggplot object", {
-     int <- sw_combine(fellingdateR:::dummy2, plot = FALSE)
+     int <- sw_combine(dummy2, plot = FALSE)
      testthat::expect_true("ggproto" %in% class(sw_combine_plot(int)$layers[[1]]$stat))
 })
