+++ tests/testthat/test-read_fh.R
@@ -1,14 +1,11 @@
 testthat::test_that("Output is data.frame", {
-
-     x <- system.file("extdata", "DOEL1.fh", package = "fellingdateR")
-     x <- read_fh(x, verbose = TRUE, header = FALSE)
-     testthat::expect_s3_class(x, "data.frame")
+        x <- system.file("extdata", "DOEL1.fh", package = "fellingdateR")
+        x <- read_fh(x, verbose = TRUE, header = FALSE)
+        testthat::expect_s3_class(x, "data.frame")
 })
 
 testthat::test_that("Output is data.frame with 29 columns", {
-
-     x <- system.file("extdata", "DOEL1.fh", package = "fellingdateR")
-     x <- read_fh(x, verbose = FALSE, header = TRUE)
-     testthat::expect_equal(ncol(x), 29)
+        x <- system.file("extdata", "DOEL1.fh", package = "fellingdateR")
+        x <- read_fh(x, verbose = FALSE, header = TRUE)
+        testthat::expect_equal(ncol(x), 29)
 })
-
