+++ pkgdown.yml
@@ -1,0 +1,10 @@
+pandoc: 2.19.2
+pkgdown: 2.0.7
+pkgdown_sha: ~
+articles:
+  getting_started: getting_started.html
+last_built: 2023-11-28T15:37Z
+urls:
+  reference: https://hanecakr.github.io/fellingdateR/reference
+  article: https://hanecakr.github.io/fellingdateR/articles
+
