+++ R/sw_sum_plot.r
@@ -15,9 +15,9 @@
 #' @export
 #'
 #' @examples
-#' dummy6 <- fellingdateR:::dummy6
-#' dummy6
-#' tmp <- sw_sum(dummy6)
+#' trs_example6 <- fellingdateR:::trs_example6
+#' trs_example6
+#' tmp <- sw_sum(trs_example6)
 #' sw_sum_plot(tmp, bar_col = "burlywood1")
 #'
 sw_sum_plot <- function(x,

[…]
