+++ R/data.R
@@ -13,8 +13,8 @@
 #' @details
 #'     sample size = 69 observations
 #' @examples
-#' sw_data_overview()
 #' sw_data_info("Brathen_1982")
+#'
 #' sw_model("Brathen_1982")
 #'
 "Brathen_1982"
@@ -35,8 +35,8 @@
 #' @details
 #'     sample size = 490 observations
 #' @examples
-#' sw_data_overview()
 #' sw_data_info("Hollstein_1980")
+#'
 #' sw_model("Hollstein_1980")
 #'
 "Hollstein_1980"
@@ -55,10 +55,10 @@
 #'   }
 #' @details
 #'     sample size = 295 observations
-#' @source Miles D. 1997. The interpretation, presentation and use of tree-ring dates. Vernacular architecture 28, 40–56. DOI: https://doi.org/10.1179/030554797786050563.
+#' @source Miles D. 1997. The interpretation, presentation and use of tree-ring dates. Vernacular architecture 28, 40–56. <https://doi.org/10.1179/030554797786050563>
 #' @examples
-#' sw_data_overview()
 #' sw_data_info("Miles_1997_NM")
+#'
 #' sw_model("Miles_1997_NM")
 #'
 "Miles_1997_NM"
@@ -77,10 +77,10 @@
 #'   }
 #' @details
 #'     sample size = 406 observations
-#' @source Miles D. 1997. The interpretation, presentation and use of tree-ring dates. Vernacular architecture 28, 40–56. DOI: https://doi.org/10.1179/030554797786050563.
+#' @source Miles D. 1997. The interpretation, presentation and use of tree-ring dates. Vernacular architecture 28, 40–56. <https://doi.org/10.1179/030554797786050563>
 #' @examples
-#' sw_data_overview()
 #' sw_data_info("Miles_1997_SC")
+#'
 #' sw_model("Miles_1997_SC")
 #'
 "Miles_1997_SC"
@@ -99,10 +99,10 @@
 #'   }
 #' @details
 #'     sample size = 219 observations
-#' @source Miles D. 1997. The interpretation, presentation and use of tree-ring dates. Vernacular architecture 28, 40–56. DOI: https://doi.org/10.1179/030554797786050563.
+#' @source Miles D. 1997. The interpretation, presentation and use of tree-ring dates. Vernacular architecture 28, 40–56. <https://doi.org/10.1179/030554797786050563>
 #' @examples
-#' sw_data_overview()
 #' sw_data_info("Miles_1997_WBC")
+#'
 #' sw_model("Miles_1997_WBC")
 #'
 "Miles_1997_WBC"
@@ -123,8 +123,8 @@
 #'     sample size = 219 observations
 #' @source Pilcher J.R. 1987. A 700 year dating chronology for northern France. Applications of tree-ring studies. Current research in dendrochronology and related subjects. BAR International Series 333, 127–139.
 #' @examples
-#' sw_data_overview()
 #' sw_data_info("Pilcher_1987")
+#'
 #' sw_model("Pilcher_1987")
 #'
 "Pilcher_1987"
@@ -143,11 +143,11 @@
 #'   }
 #' @details
 #'     sample size = 562 observations
-#'     !!! sapwood determined by *color*
[…]
