+++ man/Brathen_1982.Rd
@@ -25,8 +25,8 @@ Sapwood data set for Western Sweden published by Bräthen in 1982.
 }\if{html}{\out{</div>}}
 }
 \examples{
-sw_data_overview()
 sw_data_info("Brathen_1982")
+
 sw_model("Brathen_1982")
 
 }
