+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   getting_started: getting_started.html
-last_built: 2023-11-28T18:39Z
+last_built: 2023-11-28T18:46Z
 urls:
   reference: https://hanecakr.github.io/fellingdateR/reference
   article: https://hanecakr.github.io/fellingdateR/articles
