### 2023-11-28 #618 @hanecakr
Dear @maelle. I've been working on the package and think I'm able now to address most of your comments.
Thanks a lot for your time, valuable insights and recommendations!

I'll list the answers to your comments here:

### Installation instructions
For brevity you can keep only the installation instructions with pak.

- [x ] Ok. I've removed the install with devtools from README

### Dependencies
I wonder whether the dependency on plyr for very few functions could be avoided?

- [x ]	The function `plyr::round_any` is now removed from `sw_interval_plot()`, `sw_combine_plot()` and `sw_sum_plot()`
- [x ]  `floor/ceiling(x/10)*10 ` now does the job

### Metadata
I don't think it's allowed to have backticks in a DESCRIPTION file, I see some in the Description field. Or if it's allowed, at least it's unusual?

- [x ] Backticks removed from DESCRIPTION

### Docs
I'd recommend creating a [pkgdown website](https://devguide.ropensci.org/building.html#website).

- [x ] Initiated a pkgdown website now. I had postponed this because I also have a short paper ready (after code-review) for JOSS. Planned to use that paper (and reviewer comments) for an article/vignette on a pkgdown site. But obviously, for reviewers a pkgdown-site is most helpful at this stage already.

### Data
I see the example datasets are saved as internal datasets however you use them in the documentation (in the README) so why not make them exported datasets? https://r-pkgs.org/data.html#sec-data-data

- [x ] The exported data are factual (published) data that are used by nearly all functions in the package. With `sw_data_overview() `users get an overview of what sapwood datasets are available in the package and with `sw_data_info()` basic descriptives of those dataset.
- [ ] When I would make the made-up date also external, that would mix real-world data and made-up data. And would also complicate the code for `sw_data_overview()`. Therefore I would like to keep the made-up data for examples and tests as internal data.

Rather than using the names with "dummy" I'd recommend informative names. (Beside, some might consider "dummy" to be an offensive word related to ableism).

- [x ] Thanks for pointing this out. Changed this throughout all functions and tests to “trs_example1” (tree-ring series example). Indeed much better than ‘dummy’ or ‘fake’.

### Tests
I see some top-level code in for instance https://github.com/hanecakr/fellingdateR/blob/master/tests/testthat/test-fd_report.R I'd recommend instead having a function defined in a helper file like tests/testthat/helper-testdata
...

- [x ] The top-level code is in fact redundant. The test can equally be run with the internal data. This has now been implemented in the tests.

I think some test lines could be more condensed, e.g. ...

- [x ] I reformatted most test-files to make them more condensed. Should have done that before submission…

You shouldn't namespace fellingdateR::: data in test files.

- [x ] Right. Removed the namespace to fellingdateR in tests.

For test-sw_combine_plot.R you might want to look into snapshot testing rather than using ggplot2's internal data structure: https://testthat.r-lib.org/articles/snapshotting.html#whole-file-snapshotting (other relevant source: https://www.tidyverse.org/blog/2022/09/playing-on-the-same-team-as-your-dependecy/#testing-testing)

- [ ]  I've left this as it is now. Can't find an elegant way to implement snapshot testing right now. Need to look deeper into that.

Why is there a test that is commented out? https://github.com/hanecakr/fellingdateR/blob/05c449fb13b678ac1700177d73a20fad6c55da8c/tests/testthat/test-sw_interval.R#L2 (as opposed to, say, completely removed)

- [x ] Oops, Removed this from the test-file

Instead of the regular expressions for testing errors, you might be interested in this approach of classed errors: https://www.mm218.dev/posts/2023-11-07-classed-errors/

### Code
Please read the output from lintr in the automatic checks above, and fix accordingly (or update this thread to explain why you disagree).

Lines such as https://github.com/hanecakr/fellingdateR/blob/05c449fb13b678ac1700177d73a20fad6c55da8c/R/cor_table.R#L117 should use inherits()
if (!inherits(x, "rwl")) {

- [x ] Now using inherits() in the latest version of `cor_table()`.

In for instance cor_table.R, I think the code might be more readable with explaining variables. For instance
if (!all(diff(as.numeric(row.names(x))) == 1)) {
would be
increasing_consecutive_years <- (all(diff(as.numeric(row.names(x))) == 1))
if (!increasing_consecutive_years) {

- [x ] OK, I've changed this to increase readability

Same comments as for test files, I think there could be fewer new lines in scripts such as https://github.com/hanecakr/fellingdateR/blob/master/R/cor_table.R, to facilitate reading / vertical scrolling. I guess this conflicts a bit with the huge indentations in the files (which you are obviously allowed to prefer!).

- [x ] I’ve reformatted the code with RStudio > reformat code. Is indeed (a bit) more condensed now.

For the else if in https://github.com/hanecakr/fellingdateR/blob/05c449fb13b678ac1700177d73a20fad6c55da8c/R/sw_interval.R#L236, why not use the switch() function? I must confess I like using it but have to look up the docs every time to remember how to input the arguments. But it'll decrease the complexity of the current logic.

- [ ] I remember I tried this before, but failed to implement `switch()` for this function. Would prefer to leave it as it is now.

In https://github.com/hanecakr/fellingdateR/blob/05c449fb13b678ac1700177d73a20fad6c55da8c/R/sw_interval_plot.R#L137 what warnings are suppressed? In the same post mentioned previously https://www.mm218.dev/posts/2023-11-07-classed-errors/, it's shown that one can selectively suppress warnings of a given class.

- [x ] Doesn’t seem to be necessary anymore to use `suppressWarnings() `to avoid messages from ggplot, for `sw_model() `and `sw_combine()`. So I’ve removed the calls to `suppressWarnings() `for these functions.
- [ ] For sw_interval() however I’ve kept the call to suppressWarnings to avoid warnings as:

	Warning: Removed 52 rows containing non-finite values (`stat_align()`).
	Warning: Removed 43 rows containing missing values (`geom_line()`).

I'm sure there are more elegant and effective ways to improve the scripts, functions and their documentation, but I hope that these initial revisions have elevated the quality of the package sufficiently, making it amenable for the review process. Of course, I'm open for additional insights and recommendations to further improved the quality and performance of the package.

-- Kristof
