### 2023-11-30 #618 @hanecakr
@maelle Thanks a lot!
Meanwhile, I will change the branch name to `main`, use the `data(...)` function to load internal data, and have a look at the  .Rproj.user folder and try to find a way to add it to .gitignore. The badge will be added to README soon.
