### 2023-11-30 d8a6a2a
add badge rOpenSci
