+++ README.Rmd
@@ -20,6 +20,7 @@ knitr::opts_chunk$set(
 [![R-CMD-check](https://github.com/hanecakr/fellingDateR/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/hanecakr/fellingDateR/actions/workflows/R-CMD-check.yaml)
 [![Codecov test coverage](https://codecov.io/gh/hanecakr/fellingDateR/branch/master/graph/badge.svg)](https://app.codecov.io/gh/hanecakr/fellingDateR?branch=master)
 [![pkgcheck](https://github.com/hanecakr/fellingDateR/workflows/pkgcheck/badge.svg)](https://github.com/hanecakr/fellingDateR/actions?query=workflow%3Apkgcheck)
+[![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/618_status.svg)](https://github.com/ropensci/software-review/issues/618)
 <!-- badges: end -->
 
 This R-package offers a set of functions that can assist you to infer felling date estimates from dated tree-ring series. The presence of (partially) preserved sapwood or waney edge allows to estimate a range for the actual felling date, for individual series as well as for a group of timbers. Furthermore, an additional function provides a tool to sum sapwood probability distributions, comparable to 'summed probability densities' commonly applied to sets of radiocarbon (^14^C) dates.
@@ -67,9 +68,9 @@ sw_model("Hollstein_1980")
 
 ## Getting started
 
-An overview of the main functions and sapwood datasets included within the package can be found [here](https://hanecakr.github.io/fellingdateR/reference/).
+You can find an overview of the **main functions** and **sapwood datasets** distributed with the package in the [documentation](https://hanecakr.github.io/fellingdateR/reference/).
 
-Some examples are provided in the [`getting started`](https://hanecakr.github.io/fellingdateR/articles/getting_started.html) vignette.
+Some practical examples are provided in the [`getting started`](https://hanecakr.github.io/fellingdateR/articles/getting_started.html) vignette.
 
 
 ## Motivation
