### 2023-12-01 d542273
move internal (example) data to external data and small change to sw_data_overview()
