+++ data-raw/example_data.R
@@ -68,4 +68,4 @@ usethis::use_data(trs_example0,
                   trs_example5,
                   trs_example6,
                   trs_example7,
-                  internal = TRUE, overwrite = TRUE)
+                  internal = FALSE, overwrite = TRUE)
