+++ man/trs_example4.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{trs_example4}
+\alias{trs_example4}
+\title{Example dataset 4}
+\format{
+A tibble of 4 variables:
+\describe{
+\item{series}{unique ID of the tree-ring series}
+\item{last}{calendar year assigned to the last measured ring}
+\item{n_sapwood}{number of observed sapwood rings}
+\item{waneyedge}{waney edge present TRUE/FALSE}
+}
+}
+\usage{
+trs_example4
+}
+\description{
+A combination of series with and without sapwood rings.
+}
+\keyword{datasets}
