### 2023-12-01 #618 @hanecakr
@njtierney: with a little delay, I decided to follow the advice of @maelle and moved the example datasets from internal to external data, so they become easily available for testing and examples, both for end-uses as during code review. The latest commit implements the necessary (small) changes.  Looking forward to your comments and suggestions. -- Kristof.
