### 2023-12-31 #618 @ajpelu
Sorry for the slight delay in the review. Here it is.

I think the target audience of the pkg would be indicated more explicitly.

The images are not very illustrative of what the package does, they simply seem to indicate that they are wooden tokens. Perhaps it would be interesting to think of a schematic or similar, but the photos do not seem to provide much.

In the installation instructions, it might be more effective to provide a link to the GitHub repository of the package rather than displaying the raw GitHub link itself.

Why does the author prefer to use the package pak rather than devtools to install the development version of the pkg? I see that in the vignette he indicated the two ways (pak and devtools).

The vignette outlines the primary functions but lacks a straightforward workflow for beginners.

There are missing citations in the vignette. Please add a reference list at the end.

In the vignette of the website it appears:

```
sw_sum(fellingdateR:::trs_example7, plot = TRUE)

```
but when I run locally i got:
```
sw_sum(fellingdateR:::trs_example7, plot = TRUE)
Error in as.data.frame(x) : object 'trs_example7' not found
```
Please consider change the three `:::` by `::`

No community guidelines found, only in DESCRIPTION there is a BugReports linked.

All test pass. The coverage is 77.44 % (using devtools::test_coverage())

  - Please consider to use lowercase name as indicated by 1.1.1 point of the packaging guidelines

  - Not all the functions follow the object_verb() naming scheme. For instance, the functions read_fh and get_header. I would suggest to author rename them to fh_read and fh_header

  - The movAv function's name deviates from the consistent use of snake_case observed in the rest of the functions. To maintain uniformity, it would be preferable to rename it in accordance with the snake_case as indicated in the packaging guidelines

  - In the get_header, movAv, read_fh functions please consider change cat() or print() for message or warning, as packaging guidelines indicated.

---

Thank you for your contribution with the fellingdateR package. It provides a valuable suite of functions for estimating, reporting, and combining felling dates of historical tree-ring series. The comprehensive set of tools makes it easier to infer felling date estimates, considering factors like preserved sapwood or waney edge.

Given my expertise and interest in dendrochronology, I thoroughly enjoyed reviewing this package. The functions enable users to estimate felling date ranges for both individual series and groups of timbers, offering a valuable feature for summing sapwood probability distributions.

This being my first software review for Ropensci, I apologize for any potential errors or unclear comments. While I've provided various comments (a lot of them), feel free to focus on those that are most relevant or useful. I commend you for effectively addressing a challenging problem with this package. If you need clarification or assistance implementing the suggestions, please reach out. Thank you for submitting this software; the review process was a pleasure!

- The data description lacks a consistent pattern. The author occasionally employs country abbreviations while at other times refrains from doing so. Let's strive for coherence in this regard.

- Some code and documentation overpass the 80 characters (particularly data.R). Please consider adjust to this recommended length.

- The README file does not contain any information about how to cite. I found that one in the CITATION file and also in the website of the package. According to Ropensci packaging guidelines the README need to include information about how to cite.

- Please consider generate a theme for all plots of your package in a utils.R function

- I found very useful the /paper/paper.md, and I think much of this information would improve also the documentation and the website of the package. For instance, the fellingdateR_workflow is very useful to understand how the package works. Would the author consider include it in the REAMDE or in the DESCRIPTION of the package instead of the current pictures? I think the former is more informative than the latter

- Why is there no auxiliary function to sw_model (eg. sw_model_plot) like the rest of the functions in the package (eg. sw_interval and sw_interval_plot)? This could lighten the function size and also would be coherent with the rest of the package.

- The code style isn't consistent. Please consider use the styler package.

- I didn't find top-level documentation:
```
?fellingdateR
#No documentation for ‘fellingdateR’ in specified packages and libraries:
#you could try ‘??fellingdateR’```
Pleas consider generate ?fellingdateR or `?fellingdateR-package`. See  https://devguide.ropensci.org/building.html#general
```
- The package has a well-structured website, however I thin much of the information included in the paper.md is not in the website. For instance the background and Statement of need within the paper.md is very useful and I think it would help to the potential users of the pacakge. Please consider include in the README or in an specific vignette.

- The vignette "getting started" includes some citations (eg. Hollstein 1980) but not a full reference for this citations. Please include them at the end of the vignette.

- Please be consistent with the use of = or <-, but avoid mixed them. For instance in hdi.R

- **spell check**: DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'. Please consider to specify.

- Is there **code duplication** in the package that should be reduced?
Yes for instance the ggthemes within the plot functions

### Coverage
Please consider review the test coverage of the movAv.R function (~57.14%)

### Goodpractices
After run the `goodpractices::gp()` several code lines does not pass the check:

  ✖ write short and simple functions. These functions have high
    cyclomatic complexity (>50): read_fh (150). You can make them
    easier to reason about by encapsulating distinct steps of your
    function into subfunctions.
  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    R/cor_table.R:108:18
    R/cor_table.R:109:22
    R/cor_table.R:112:22
    R/hdi.R:48:12
    R/hdi.R:49:14

    data-raw/DATASETS.Rmd:17:81
    data-raw/DATASETS.Rmd:19:81
    data-raw/DATASETS.Rmd:35:81
    data-raw/DATASETS.Rmd:51:81
    data-raw/DATASETS.Rmd:68:81

    R/cor_table.R:192:20
    R/cor_table.R:195:20

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...)
    and 1:NCOL(...) expressions. They are error prone and result 1:0 if
    the expression on the right hand side is zero. Use seq_len() or
    seq_along() instead.

    R/fd_report.R:162:19
    R/hdi.R:54:16
    R/sw_combine.R:123:27
    R/sw_combine.R:180:27
    R/sw_combine.R:361:35

  ✖ fix this R CMD check NOTE: Note: found 11 marked UTF-8
    strings

  ✖ avoid 'T' and 'F', as they are just variables which are set
    to the logicals 'TRUE' and 'FALSE' by default, but are not reserved
    words and hence can be overwritten by the user.  Hence, one should
    always use 'TRUE' and 'FALSE' for the logicals.

    R/sw_combine_plot.R:NA:NA
    R/sw_combine_plot.R:NA:NA
    R/sw_combine_plot.R:NA:NA
    R/sw_combine_plot.R:NA:NA
    R/sw_combine_plot.R:NA:NA

#### `read_fh.R`

- Whats new funtionalities include read_fh not covered by dplR::read.fh()?
- There are several code commented that could be removed. eg.
    `# NEW: verbose = TRUE, header = FALSE`
    Please revise them
- The nomenclature of the objetcs exhibits a mix of snake_case and dot.case. Let's ensure consistency in the naming convention.
- The function appears to be extensive. Has the author explored the possibility of breaking it down into several auxiliary functions to enhance readability and maintainability?
- In the function documentation, the author mentions the Heidelberg format file. It would be beneficial to include a link to the specifications of this format file for reference, eg. https://www.treeringsociety.org/resources/SOM/Brewer_Murphy_SupplementaryMaterial.pdf

#### `get_header.R`

- Please consider include some defensive programming such us:

```
if (!is.data.frame(rwl) || !inherits(rwl, "rwl")) {
    stop("Input should be a data.frame of class 'rwl'")
  }
```

- avoid `cat` as indicated in packaging guidelines

#### `cor_table.R`

- Very interesting and useful function. Has the author contemplated breaking down this function into more modular components? For instance, each of the correlation values could potentially be extracted as auxiliary functions, providing users with the flexibility to call them individually as needed.

- The nomenclature of arguments exhibits a mix of snake_case and dot.case. Let's ensure consistency in the naming convention.
- Remove code comments in the argument section, eg.
```
...
output = "table",
#c("matrix", "table")
...
```
- In the documentation for the argument `sort_by`, it would be useful to add the choices to avoid errors. For instance,
```
#' @param sort_by The correlation value by which the output is sorted for each
#'   series in `x`. One of "r_pearson", "t_St", "glk", "glk_p", "t_BP", "t_Ho". Default to "t_Ho"
```

- Ensure uniformity in assignment by using the same operator consistently—avoid mixing `<-` and `=`

- Consider avoid the use of sapply. See this https://adv-r.hadley.nz/functionals.html. For instance, consider
```
y <- y[, apply(!is.na(y), 2, sum) > 3, drop = FALSE]
```
instead of
```
y <-
               y[, sapply(y, function(col)
                    sum(!is.na(col))) > 3, drop = FALSE]
```

#### `sw_model.R`

- Ensure uniformity in assignment by using the same operator consistently—avoid mixing `<-` and `=`. eg. ` n_obs = sum(observed$count)`

- I'm not sure if incorporating the reading of a csv is optimal in a function. What if a user wants to input a dataframe that isn't from a CSV file? I propose modifying the function to accept a dataframe or tibble with mandatory `n_sapwood` and `count` columns. Additionally, allowing users to import from a CSV could be achieved by either creating a small auxiliary function or simply utilizing utils::read.csv(). This also reduce the size of the sw_model function. Please consider this for all those functions with the same issue.

- I think it is better to have a separate function for plotting, which could be called after the computation is completed by sw_model, for instance, sw_model_plot. This also allow to user modify the plot generated. For instance if the user doesn`t like the grid within the plot. In the current form to do that the user need to type:

```
s <- sw_model("Hollstein_1980", plot  = TRUE)
s + theme(panel.grid = element_blank())
```

So if there is an specific function to plot the data, this function would include the `...` in the arguments, to allow include other ggplot2 arguments.

- Anycase, in the plot section would be better to use .data, for instance, in
```
ggplot2::geom_segment(
  data = hdi_model,
  ggplot2::aes(
    x = hdi_model[[1]],
    xend = hdi_model[[2]],
    y = 0,
    yend = 0
  ),
  colour = "grey30",
  linewidth = .8,
  alpha = 0.8
)
```

I suggest to use `.data[[1]]`

```
ggplot2::geom_segment(
  data = hdi_model,
  ggplot2::aes(
    x = .data[[1]],
    xend = .data[[2]],
    y = 0,
    yend = 0
  ),
  colour = "grey30",
  linewidth = .8,
  alpha = 0.8
)
```
- In the documentation of sw_model, the values of the densfun parameter are not in italics and are enclosed in double quotation marks. It doesn't matter, but please maintain consistency with the rest of the documentation. For example, in sw_interval.R, they are in italics.

#### `sw_interval.R`

- Please remove code commented

        # if (last != 0) {

- In this function, the separation of the plot part from the computation one is appreciated. A further step would be to remove the plotting option completely from the function, and specified in the documentation (and also in the vignette) to use the sw_interval_plot function to obtain the plot.

#### `sw_interval_plot.R`

- I would suggest to add the ellipsis argument (i.e. `...`) to allow another ggplot2 arguments. See also [ellipsis](https://ellipsis.r-lib.org/) pkg for more info. Consider apply this throughout the plot functions of the package.

- Which notes are the author trying to avoid with?
```
# to avoid notes in CMD check
        p.x <- upper <- year <- NULL
```
¿any alternative to solve them?

- Please, consider change this code
```
        if (all(
                !(attributes(x)$names) %in% c(
                        "year",
                        "n_sapwood",
                        "p")
        ))
        stop("Input differs from output sw_interval()")
```
by:

```
if (!all(c("year", "n_sapwood", "p") %in% names(attributes(x)))) {
  stop("Input structure differs from the expected output of sw_interval()")
}

```

- You mixed snake_case with dot.case in the naming of objects. I would suggest to choose one and be consistent.

- What this code snippet do? If I understand correctly, the sw_interval_plot requires the output of sw_interval, so is a .csv format an output of the sw_interval?
```
        if (grepl("\\.csv$", sw_data.p))
                sw_data.p <- basename(sw_data.p)
```

- Please specify the meaning of the 'p' on the y-axis.

#### `sw_combine.R`

- This function requires a dataframe, as indicated in the documentation. However, there is no check for that; instead, the first line forces x to be a data.frame. I would suggest to add a defensive programming line such as:
```
   if (!is.data.frame(x)) {
    stop("Input 'x' must be a dataframe.")
  }
```

- Camel case and snake_case styles are mixed in the function. Please be consistent.
- There are some duplications in the function, for instance `pdf_matrix[, 1] <- timeAxis`
- Please revise the comments within the function. Are they all necessary? Example:
```
      # when multiple exact felling dates are listed that do no correspond
      # --> COMB = 0 and after scaling NaN (division by 0)
      # check rowwise if there is any p-value == 1,
      # and replace COMB at that position with 1
```

- Perhaps this function could be modularized into several simpler functions. Initially, a function could evaluate whether the dataset contains all series with preserved sapwood or if any series has an exact felling date, among other conditions (as indicated in the examples of the function). Subsequently, based on the initial evaluation, the function could call auxiliary functions.

- In any case, the function should provide detailed information about the approaches used for each case.

- Please, see my comment about the data input in `csv` format for the sw_model.R

- The elements of the output list generated by this function are not described. Please add to the documentation.

#### `sw_combine_plot.R`

- Please consider taking out the rescale function out of the sw_combine_plot function.
- A lot of mixing T, F with TRUE and FALSE. Please be consistent.
- Camel case and snake_case styles are mixed in the function. Please be consistent.
- I would suggest to allow the user customize the color used. See my comment in the below section (Plot functions).

- Remove the commented codes inside the function:
```
      # NEXT LINE TRIGGERS WARNING
      # Warning message:
      # Using one column matrices in `filter()` was deprecated in dplyr 1.1.0.
      # ℹ Please use one dimensional logical vectors instead.
      # ℹ The deprecated feature was likely used in the fellingdateR package.
      # Please report the issue to the authors.
      # { if (nrow(summary |> dplyr::filter(agreement == "poor")) != 0)
      # replaced by:
```

- The function lacks an axis legend. I presume that the x-axis corresponds to calendar time, and the y-axis corresponds to 'p'. Is my understanding correct?

- Any way to indicate the "A_i" as real subscript? maybe parse = TRUE within geom_text?

- Please add information in the documentation about the agregation index
- In the second example:
```
combo <- fellingdateR::sw_combine(trs_example2)
fellingdateR::sw_combine_plot(combo)
```

- What do the arrow and the dot refer? Please include in the documentation.

#### `sw_sum.R`
- In the documentation there is a mistake. In the description of the `last` argument please add a blank space in `xcontaining`
- In the documentation the values of the densfun parameter are not in italics and are enclosed in double quotation marks. It doesn't matter, but please maintain consistency with the rest of the documentation. For example, in sw_interval.R, they are in italics.

- Regarding defensive programming idem comment that for sw_combine. This function requires a dataframe, as indicated in the documentation. However, there is no check for that; instead, the first line forces x to be a data.frame.  I would suggest to add a defensive programming line such as:
```
   if (!is.data.frame(x)) {
    stop("Input 'x' must be a dataframe.")
  }
```

#### `sw_sum_plot.R`
- Good point here using customizable colors. Let's implement this feature consistently across other plot functions for a cohesive and customizable user experience.

- Can a user change the width of the moving average window used within the sw_sum_plot()? It seems that the sw_sum_plot() uses always the same.

- Idem with the splines used.

- What do the blue dots, blue bars and red line mean? Please indicate in the documentation

### Other comments

- I am surprised that different functions (for instance sw_combine and sw_sum), that accessing similar subdata within a data set, use different approaches:
```
cambium <- df[[waneyedge]]
```
and
```
cambium <- x[, waneyedge]
```

Please be consistent along the package.

- Please review the naming conventions, including CamelCase, snake_case, and dot.case, used for objects throughout the package. It is recommended to choose a consistent style, and I would suggest using snake_case. Ensure uniformity across the package for improved clarity and maintainability.

- I'm not sure if incorporating the reading of a csv is optimal in a function. What if a user wants to input a dataframe that isn't from a CSV file? I propose modifying the function to accept a dataframe or tibble with mandatory `n_sapwood` and `count` columns. Additionally, allowing users to import from a CSV could be achieved by either creating a small auxiliary function or simply utilizing utils::read.csv(). This also reduce the size of the sw_model function.

- In the different plot functions, the authors uses a common theme. It might be useful to create an auxiliary function with a ggplot2::theme() specification. Then, each plot function could used this auxiliary function by default or also allow to the user to specify a custom theme.

- I would suggest changing the specified colors in each plot function to custom arguments, allowing users to choose their preferred colors. For instance, in th sw_interval_plot the colors of the area specified by

```
                ggplot2::geom_area(
                        ggplot2::aes(x = ifelse(year >= lower, year, NA),
                                     y = p.x),
                        fill = "tomato3",
                        color = "tomato3",
                        alpha = 0.3,
                        linewidth = 1
```
could be indicated as arguments, e.g. area_fill and area_color.
