### 2024-01-02 #618 @hanecakr
Thank you for your time investment and extensive review report. I hope to find some time next week to start working on all comments and suggestions. Might need some further assistance to tackle all issues raised.
Happy new year! 🎇🎇🎇
