### 2024-02-09 #618 @hanecakr
Dear Maëlle,

I very sorry for the horrible delay in my response to the review reports. Especially given the timely reports of the reviewers.
Every year, jan-feb are months at work where most of my time is consumed by strickt deadlines.
But in a week or two, sky’s look bright again 😉 and I will start working on the review!

have a nice day,
Kristof.

Kristof Haneca
Erfgoedonderzoeker | dendrochronoloog

Agentschap Onroerend Erfgoed
M +32 (0)474 44 81 96
***@***.***<https://orcid.org/0000-0002-7719-8305>  ***@***.*** <https://twitter.com/KristofHaneca>   ***@***.*** <https://www.researchgate.net/profile/Kristof_Haneca>

From: Maëlle Salmon ***@***.***>
Sent: Friday, February 9, 2024 7:54 AM
To: ropensci/software-review ***@***.***>
Cc: Haneca Kristof ***@***.***>; Mention ***@***.***>
Subject: Re: [ropensci/software-review] fellingdateR: Estimate, report and combine felling dates of historical tree-ring series (Issue #618)

👋 @hanecakr<https://github.com/hanecakr>, any update? 😸
