### 2024-02-24 8bbd7c4
initiating new branch and first edits according to rOpenSci review
