+++ R/get_header.R
@@ -17,20 +17,18 @@
 #' @export
 
 get_header <-
-     function(rwl) {
-          if (identical(class(rwl), c("rwl", "data.frame"))) {
-               attr(rwl, "row.names") <- NULL
-               attr(rwl, "po") <- NULL
-               attr(rwl, "class") <- NULL
-               attr(rwl, "names") <- NULL
+  function(rwl) {
+    if (identical(class(rwl), c("rwl", "data.frame"))) {
+      attr(rwl, "row.names") <- NULL
+      attr(rwl, "po") <- NULL
+      attr(rwl, "class") <- NULL
+      attr(rwl, "names") <- NULL
 
-               tmp <- attributes(rwl)
-               tmp <- data.frame(tmp)
+      tmp <- attributes(rwl)
+      tmp <- data.frame(tmp)
 
-               return(tmp)
-
-          } else {
-               cat(gettext('Object should be a data.frame of class "rwl"\n'))
-
-          }
-     }
+      return(tmp)
+    } else {
+      cat(gettext('Object should be a data.frame of class "rwl"\n'))
+    }
+  }
