+++ R/cor_table.R
@@ -17,15 +17,12 @@
 #'   series and references are removed from the output.
 #' @param output The desired output format, either "matrix" or "table"
 #'   (default).
-#' @param values A character vector listing correlation values to be computed.
-#'   Defaults to all possible values: c("r_pearson", "t_St", "glk", "glk_p",
-#'   "t_BP", "t_Ho").
 #' @param sort_by The correlation value by which the output is sorted for each
 #'   series in `x`.
 #'
 #' @return The function returns a list of correlation matrices if `output` is
 #'   set to "matrix." If `output` is set to "table," it returns a `data.frame`
-#'   containing correlation values listed by the specified `values`.
+#'   reporting all correlation values.
 #' @export
 #' @references
 #'   - Baillie, M.G.L., Pilcher, J.R. (1973) A simple crossdating program for
@@ -88,374 +85,324 @@
 #' cor_results <- cor_table(Doel1_trs, output = "table", min_overlap = 80)
 #'
 cor_table <-
-     function(x,
-              y = NULL,
-              min_overlap = 50,
-              remove.duplicates = TRUE,
-              output = "table",
-              #c("matrix", "table")
-              values = c("r_pearson", "t_St", "glk", "glk_p", "t_BP", "t_Ho"),
-              sort_by = "t_Ho")
-          #c("glk", "t_St", "t_BP", "t_Ho", "r_pearson"))
-     {
-          # to avoid notes in CMD check
-          pnorm <- cor <- select <- series <- reference <- NULL
+  function(x,
+           y = NULL,
+           min_overlap = 50,
+           remove.duplicates = TRUE,
+           output = "table", # c("matrix", "table")
+           sort_by = "t_Ho") # c("glk", "t_St", "t_BP", "t_Ho", "r_pearson"))
+  {
+    # to avoid notes in CMD check
+    pnorm <- cor <- select <- series <- reference <- NULL
 
-          ### checks
-          x_ori <- x
+    ### checks
+    x_ori <- x
 
-          if (is.null(y)) {
-               y = x
-               noRef = TRUE
-          }
-          else {
-               noRef = FALSE
-               y_ori <- y
+    if (is.null(y)) {
+      y <- x
+      noRef <- TRUE
+    } else {
+      noRef <- FALSE
+      y_ori <- y
+    }
 
-          }
+    if (!inherits(x, "rwl")) {
+      warning("'x' is not class rwl")
+    }
+    if (!inherits(y, "rwl")) {
+      warning("'y' is not class rwl")
+    }
 
-          if (!inherits(x, "rwl"))  {
-               warning("'x' is not class rwl")
-          }
-          if (!inherits(y, "rwl"))  {
-               warning("'y' is not class rwl")
-          }
+    increasing_consecutive_years <- all(diff(as.numeric(row.names(x))) == 1)
+    if (!increasing_consecutive_years) {
+      stop(
+        "rownames 'x' are not consecutive years in increasing order."
+      )
+    }
 
-          increasing_consecutive_years <- all(diff(as.numeric(row.names(x))) == 1)
-          if (!increasing_consecutive_years) {
-               stop(
-                    "The tree-ring series 'x' have/has no consecutive years in increasing order as rownames."
-               )
-          }
-
-          increasing_consecutive_years <- all(diff(as.numeric(row.names(y))) == 1)
-          if (!increasing_consecutive_years) {
-               stop(
-                    "The master series 'y' have/has no consecutive years in increasing order as rownames."
-               )
-          }
-          if (any(
[…]
