+++ R/fellingdateR-package.R
@@ -1,0 +1,30 @@
+#' fellingdateR
+#'
+#' @description fellingdateR is an R package that facilitate the analysis and
+#' interpretation of tree-ring data from wooden cultural heritage objects and
+#' structures. The package standardizes the process of computing and combining
+#' felling date estimates (fd), derived from the observed number of sapwood
+#' rings (sw), both for individual and groups of related tree-ring series.
+#'
+#' @seealso
+#'
+#' It's main functions are:
+#' \itemize{
+#'   \item [sw_model()]
+#'   \item [sw_interval()]
+#'   \item [fd_report()]
+#'   \item [sw_combine()]
+#'   \item [sw_sum()]
+#' }
+#'
+#' @section Learn more about fellingdater at:
+#' \url{https://hanecakr.github.io/fellingdateR/index.html/}
+#'
+#' @name fellingdateR
+#' @docType package
+#' @keywords package
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+NULL
