+++ R/read_fh.R
@@ -4,6 +4,9 @@
 #' ring width data and returns the HEADER fields as attributes of the
 #' `data.frame`´with the measurement data.
 #'
+#' The Heidelberg file format is described in detail here:
+#' https://www.treeringsociety.org/resources/SOM/Brewer_Murphy_SupplementaryMaterial.pdf
+#'
 #' @param fname A `character` vector specifying the name of the .fh file to
 #'   read.
 #' @param BC_correction A `logical`. If `TRUE` the correction moves BC-dates one
@@ -57,1176 +60,1181 @@ read_fh <- function(fname,
                     BC_correction = FALSE,
                     verbose = TRUE,
                     header = FALSE) {
-        # NEW: verbose = TRUE, header = FALSE
-        inp <- readLines(fname, ok = TRUE, warn = FALSE)
-        # NEW: removes empty lines in .fh file
-        inp <- inp[length(inp) != 0]
-        ## Get start and end positions of headers and data blocks
-        header.begin <- grep("^HEADER:$", inp)
-        # NEW: Quadro => chrono
-        # NEW: Double => half chrono
-        header.end <-
-                grep("^DATA:(Tree|Single|HalfChrono|Double|Quadro)$",
-                     inp)
-        n <- length(header.end)
-        if (n == 0) {
-                # NEW: warning now includes Double and Quadro
-                stop('file has no data in "Tree", "Single" of "(half)Chrono" formats')
-        }
-        ## For each data block in one of the supported formats, find the
-        ## corresponding header block
-        header.taken <- logical(length(header.begin))
-        for (i in seq_len(n)) {
-                n.preceding <- sum(header.begin < header.end[i] - 1)
-                if (n.preceding == 0 || header.taken[n.preceding]) {
-                        stop("invalid file: HEADER and DATA don't match")
-                } else {
-                        header.taken[n.preceding] <- TRUE
-                }
-        }
-        if (!all(header.taken)) {
-                warning("more HEADER blocks than DATA blocks in supported formats")
-        }
-        ## For each data block in one of the supported formats, find the
-        ## following header block (or end of file)
-        data.end <- numeric(n)
-        for (i in seq_len(n - 1)) {
-                tmp <- header.begin[header.begin > header.end[i]]
-                data.end[i] <- tmp[1]
-        }
-        tmp <- header.begin[header.begin > header.end[n]]
-        if (length(tmp) > 0) {
-                data.end[n] <- tmp[1]
+  inp <- readLines(fname, ok = TRUE, warn = FALSE)
+  # removes empty lines in .fh file
+  inp <- inp[length(inp) != 0]
+  ## Get start and end positions of headers and data blocks
+  header.begin <- grep("^HEADER:$", inp)
+  header.end <-
+    grep(
+      "^DATA:(Tree|Single|HalfChrono|Double|Quadro)$",
+      inp
+    )
+  n <- length(header.end)
+  if (n == 0) {
+    stop('file has no data in "Tree", "Single" of "(half)Chrono" formats')
+  }
+  ## For each data block in one of the supported formats, find the
+  ## corresponding header block
+  header.taken <- logical(length(header.begin))
+  for (i in seq_len(n)) {
+    n.preceding <- sum(header.begin < header.end[i] - 1)
+    if (n.preceding == 0 || header.taken[n.preceding]) {
[…]
