+++ DESCRIPTION
@@ -20,6 +20,7 @@ URL: https://github.com/hanecakr/fellingdateR,
 BugReports: https://github.com/hanecakr/fellingdateR/issues
 License: MIT + file LICENSE
 Encoding: UTF-8
+Language: en-GB
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
 Depends: 
