+++ NAMESPACE
@@ -4,7 +4,7 @@ export(cor_table)
 export(fd_report)
 export(get_header)
 export(hdi)
-export(movAv)
+export(mov_av)
 export(read_fh)
 export(sw_combine)
 export(sw_combine_plot)

[…]
