### 2024-03-11 c18e5ff
new function sw_model_plot & removal of input from .csv-file in sw_model. Input can be a user-defined data.frame.
