### 2024-03-11 e499e8c
restyling sw_interval & sw_interval_plot adding optin to choose color for plotting in sw_inteval_plot
