+++ R/sw_interval.R
@@ -11,257 +11,231 @@
 #' @param hdi A `logical.` If `TRUE`, the lower and upper limit of the highest
 #'   density interval (credible interval) are returned. When `FALSE`, a matrix
 #'   is returned with scaled p-values for each number of observed sapwood rings.
-#' @param credMass  A `scalar` in the range of `[0, 1]` specifying the mass
+#' @param cred_mass  A `scalar` in the range of `[0, 1]` specifying the mass
 #'   within the credible interval (default = .954)
 #' @param sw_data The name of the sapwood data set to use for modelling. It
-#'   should be one of the data sets listed in [sw_data_overview()], or the path
-#'   to a .csv file with columns `n_sapwood` and `count`.
+#'   should be one of the data sets listed in [sw_data_overview()], or the name
+#'   of a `data.frame` with sapwood data in columns `n_sapwood` and `count`.
 #' @param densfun Name of the density function fitted to the sapwood data set.
 #'   Should be one of:
 #'   * _lognormal_ (the default value),
 #'   * _normal_,
 #'   * _weibull_,
 #'   * _gammma_.
-#' @param sep Should be "," (comma)  or ";" (semi-colon) and is used when a
-#'   sapwood data set is provided from a user-defined .csv-file.
 #' @param plot A `logical`.
 #'   * If `TRUE`, a ggplot-style graph is returned of the individual
 #'   sapwood model and estimate of the felling date range.
 #'   * If `FALSE`, a list with the numeric output of the modelling process is returned.
 #' @examples
 #' # 10 sapwood rings observed and the Wazny 1990 sapwood model:
-#' sw_interval(n_sapwood = 10,
-#'             last = 1234,
-#'             hdi = TRUE,
-#'             credMass = .95,
-#'             sw_data = "Wazny_1990",
-#'             densfun = "lognormal",
-#'             plot = FALSE)
-#'  # same example as above, but with numerical output (hdi = FALSE):
-#' sw_interval(n_sapwood = 10,
-#'             last = 1234,
-#'             hdi = FALSE,
-#'             credMass = .95,
-#'             sw_data = "Wazny_1990",
-#'             densfun = "lognormal",
-#'             plot = FALSE)
+#' sw_interval(
+#'   n_sapwood = 10,
+#'   last = 1234,
+#'   hdi = TRUE,
+#'   cred_mass = .95,
+#'   sw_data = "Wazny_1990",
+#'   densfun = "lognormal",
+#'   plot = FALSE
+#' )
+#' # same example as above, but with numerical output (hdi = FALSE):
+#' sw_interval(
+#'   n_sapwood = 10,
+#'   last = 1234,
+#'   hdi = FALSE,
+#'   cred_mass = .95,
+#'   sw_data = "Wazny_1990",
+#'   densfun = "lognormal",
+#'   plot = FALSE
+#' )
 #'
 #' @export
 #'
 #' @return Depends on the value of `hdi`.
 #'
 #'  * If `hdi = TRUE`, a `numeric vector` reporting the upper and lower limit
-#'   of the hdi (attributes provide more detail on `credMass` and the applied
+#'   of the hdi (attributes provide more detail on `cred_mass` and the applied
 #'   sapwood model (`sw_data`)).
 #'  * If `hdi = FALSE`, a `matrix` with scaled p values for each number of
 #'   observed sapwood rings. This matrix
 
-sw_interval <- function(n_sapwood = NA,
-                        last = 1,
-                        hdi = FALSE,
-                        credMass = 0.954,
-                        sw_data = "Hollstein_1980",
-                        densfun = "lognormal",
-                        sep = ";",
-                        plot = FALSE) {
-        # Check input of `n_sapwood` and `last`. Other parameters are checked by sw_model()
-        if (is.na(n_sapwood)) {
[…]
