+++ man/fh_header.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_header.R
-\name{get_header}
-\alias{get_header}
+% Please edit documentation in R/fh_header.R
+\name{fh_header}
+\alias{fh_header}
 \title{Retrieve the HEADER fields of a Heidelberg format (.fh) file}
 \usage{
-get_header(rwl)
+fh_header(rwl)
 }
 \arguments{
 \item{rwl}{The output of \code{read_fh(x, header = TRUE)}, a \code{data.frame} of class
@@ -22,6 +22,6 @@ as attributes of the \code{data.frame} with the measurement data it returns.
 \examples{
 Doel1 <- system.file("extdata", "DOEL1.fh", package = "fellingdateR")
 Doel1_trs <- read_fh(Doel1, verbose = FALSE)
-get_header(Doel1_trs)
+fh_header(Doel1_trs)
 
 }
