+++ man/mov_av.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/movAv.R
-\name{movAv}
-\alias{movAv}
+% Please edit documentation in R/mov_av.R
+\name{mov_av}
+\alias{mov_av}
 \title{Compute a running mean on a time series}
 \usage{
-movAv(x, w = 11, align = "center", edges = "fill")
+mov_av(x, w = 11, align = "center", edges = "fill")
 }
 \arguments{
-\item{x}{A \code{numeric} vector of \code{length(x)}, containing the time series data.}
+\item{x}{A \code{numeric} vector of \code{length(x)}, with the time series data.}
 
 \item{w}{The width of the moving average window. When \code{w}is even, one more
 value from the future is included.}
@@ -24,8 +24,7 @@ It can be one of:
 series. Options are:
 \itemize{
 \item \code{edges = "fill"}: The average is calculated with a decreasing number of
-values near the edges of
-the vector (default).
+values near the edges of the vector (default).
 \item \code{edges = "nofill"}: The result includes \code{NA} values when the window does
 not cover w values.
 }}
@@ -40,6 +39,6 @@ over a continuous (time) series with a specified window width.
 }
 \examples{
 num_vec <- rnorm(100)
-movAv(num_vec, w = 20)
+mov_av(num_vec, w = 20)
 
 }
