### 2024-03-19 46540d7
introducing sw_model_plot
