+++ man/sw_model_plot.Rd
@@ -1,0 +1,40 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/sw_model_plot.R
+\name{sw_model_plot}
+\alias{sw_model_plot}
+\title{Plot function for the output of \code{\link[=sw_model]{sw_model()}}}
+\usage{
+sw_model_plot(
+  x,
+  bar_fill = "steelblue3",
+  bar_color = "grey60",
+  line_color = "red3",
+  ...
+)
+}
+\arguments{
+\item{x}{Output of \code{\link[=sw_model]{sw_model()}}.}
+
+\item{bar_fill}{Fill color for the bars (original data).}
+
+\item{bar_color}{Line color for the bars (original data).}
+
+\item{line_color}{Line color for the fitted distribution.}
+}
+\value{
+A ggplot-style graph.
+}
+\description{
+Returns a ggplot-style graph of the probability density function modelled to
+a data set of observed sapwood numbers, as computed by \code{\link[=sw_model]{sw_model()}}.
+}
+\examples{
+tmp <- sw_model(
+  sw_data = "Hollstein_1980",
+  densfun = "lognormal",
+  credMass = .95,
+  plot = FALSE
+)
+sw_model_plot(tmp)
+
+}
