### 2024-03-19 4c7cd47
defensive programming in sw_model
