### 2024-03-19 fde2f29
introducing helper-functions
