### 2024-03-19 387f2bd
rewrite, style and snake_case or sw_interval and sw_interval_plot
