### 2024-03-19 384aefa
defensive programming for sw_interval_plot
