### 2024-03-19 df9e4cc
defensive programming, style and snake_case for fd_report
