+++ R/fd_report.R
@@ -31,23 +31,20 @@
 #'   to the last measured ring (should be a `numeric` vector).
 #' @param sw_data There are two options:
 #'    * A `character` string providing the name of the sapwood data set to use
-#'   for modelling. Should be one of [sw_data_overview()], or the path to a .csv
-#'   file with columns ´n_sapwood´ and ´count´. This variable will be used for
-#'   all individual series in `x`, or
+#'    for modelling. It should be one of the data sets listed in
+#'    [sw_data_overview()], or the name of a `data.frame` with sapwood data
+#'    in columns `n_sapwood` and `count`, or
 #'
 #'    * the name of the column in `x`that lists for each series one of the
-#'   sapwood data sets given by [sw_data_overview()], or the path to a .csv file
-#'   with columns ´n_sapwood´ and ´count´.
-#' @param credMass A `scalar [0, 1]` specifying the mass within the credible
+#'   sapwood data sets given by [sw_data_overview()].
+#' @param cred_mass A `scalar [0, 1]` specifying the mass within the credible
 #'   interval (default = .954).
-#' @param densfun Name of the density function to fit to the sapwood data set.
+#' @param densfun Name of the density function fitted to the sapwood data set.
 #'   Should be one of:
-#'   * "lognormal" (the default value),
-#'   * "normal",
-#'   * "weibull",
-#'   * "gammma".
-#' @param sep Should be "," (comma)  or ";" (semi-colon) and is used when a
-#'   sapwood data set is provided from user-defined .csv-file.
+#'   * _lognormal_ (the default value),
+#'   * _normal_,
+#'   * _weibull_,
+#'   * _gammma_.
 #'
 #' @description Reports the lower and upper boundaries of a felling date range
 #'   for individual tree-ring series.
@@ -74,10 +71,10 @@
 #'                                 "vanDaalen_Norway",
 #'                                 "vanDaalen_Norway")
 #'
-#' trs_example2_edit <- cbind(trs_example2, sw_models_for_indiv_series)
+#' trs_example2_edit <- cbind(trs_example2, "sw_models" = sw_models_for_indiv_series)
 #'
 #' fd_report(trs_example2_edit,
-#'           sw_data = "sw_models_for_indiv_series"
+#'           sw_data = "sw_models"
 #'           )
 #'
 #' @export
@@ -89,74 +86,33 @@ fd_report <- function(x,
                       n_sapwood = "n_sapwood",
                       waneyedge = "waneyedge",
                       sw_data = "Hollstein_1980",
-                      credMass = 0.954,
-                      densfun = "lognormal",
-                      sep = ";") {
-        df <- as.data.frame(x)
-        # Check columns exist
-        if (!series %in% names(df)) {
-                stop("--> 'series' does not exist")
-        }
-        if (!last %in% names(df)) {
-                stop("--> 'last' does not exist")
-        }
-        if (!n_sapwood %in% names(df)) {
-                stop("--> 'n_sapwood' does not exist")
-        }
-        if (!waneyedge %in% names(df)) {
-                stop("--> 'waneyedge' does not exist")
-        }
-        series <- df[[series]] # check for NA's
-        if (any(is.na(series))) {
-                stop("--> some 'series' have no id")
-        }
-        n_sapwood <- df[[n_sapwood]] # check is.numeric
-        if (is.character(n_sapwood)) {
-                # was !is.numeric !!!
-                stop("--> 'n_sapwood' must be a numeric vector")
-        }
-        last <- df[[last]] # check is.numeric
-        if (!is.numeric(last)) {
-                stop("--> 'last' must be a numeric vector")
-        }
-        waneyedge <- df[[waneyedge]] # check is.logical
[…]
