### 2024-03-19 53c02dd
defensive programming, snake_case and code style for sw_combine and sw_combine_plot
