+++ R/sw_combine_plot.R
@@ -13,287 +13,296 @@
 #' combo <- sw_combine(trs_example2)
 #' sw_combine_plot(combo)
 #'
-#'
-sw_combine_plot <- function(x) {
-   ########################################################################
-   # function to rescale probabilities between [0, 1].
-   rescale <- function(x,
-                       floor = 0,
-                       ceiling = 1) {
-      if (max(x, na.rm = T) == 0) {
-         x
-      }
-      else{
-         (x - min(x, na.rm = T)) * (ceiling - floor) / (max(x, na.rm = T) - min(x, na.rm = T))
-      }
-   }
-   ########################################################################
+sw_combine_plot <- function(x, ...) {
 
-   if (all(
-      !(attributes(x)$names) %in% c(
-         "rawData",
-         "sapwood_data",
-         "sapwood_model",
-         "credMass",
-         "hdi_model",
-         "hdi_combine",
-         "individual_series",
-         "A_c",
-         "A_comb",
-         "model_summary"
+  if (all(
+    !(attributes(x)$names) %in% c(
+      "raw_data",
+      "sapwood_data",
+      "sapwood_model",
+      "cred_mass",
+      "hdi_model",
+      "hdi_combine",
+      "individual_series",
+      "A_c",
+      "A_model",
+      "model_summary"
+    )
+  )) {
+    stop("Input differs from output sw_combine()")
+  }
+
+  # to avoid notes in CMD check
+  year <-
+    p <-
+    lower <-
+    upper <- comb <- last <- n_sapwood <- agr_index <- agreement <- NULL
+
+
+  series <- colnames(x$raw_data)[2:(ncol(x$raw_data) - 1)]
+
+  n <- length(series)
+  cred_mass <- x$cred_mass
+  sw_data <- x$sapwood_data
+  densfun <- x$sapwood_model
+
+  pdf <- x$raw_data
+  pdf[, 2:ncol(x$raw_data)] <- apply(pdf[, 2:ncol(x$raw_data)],
+    2,
+    FUN = rescale
+  )
+  pdf <-
+    as.data.frame(pdf) # added to avoid warning by dplyr::filter
+
+  combo <- pdf[, c("year", "comb")]
+  combo <-
+    as.data.frame(combo) # added to avoid warning by dplyr::filter
+
+  range <- range(pdf$year)
+  if (length(range < 20)){range[2] <- range[2] + 20}
+
+  # creates a vector with series that only have a terminus post quem
+  tpq <- which(apply(
+    x$raw_data,
+    2,
+    FUN = function(x) {
+      sum(x, na.rm = TRUE)
+    }
+  ) == 0)
+  tpq <- names(tpq)
+  tpq <- setdiff(tpq, c("comb"))
+
+  tpq_min <- x$hdi_model[[1]]
+
+  hdi <- x$hdi_combine
+
+  summary <- x$individual_series
+  summary$agreement <-
+    ifelse(summary$agr_index >= 60 | is.na(summary$agr_index),
+      "good", "poor"
+    )
+
+  pdf <-
+    pdf |>
+    tidyr::pivot_longer(-year, names_to = "series", values_to = "p") |>
+    dplyr::filter(series != "comb") |>
+    dplyr::mutate(series = factor(series))
+
+
+  p <- ggplot2::ggplot(pdf, ggplot2::aes(x = year, y = p)) +
+
+    # plot the pdf for each series with swr or exact felling date
+    {
+      if (length(tpq) < nrow(summary)) {
+        ggplot2::geom_area(
+          fill = "grey90",
+          alpha = 0.7,
+          na.rm = T
+        )
+      }
+    } +
+    {
+      if (length(tpq) < nrow(summary)) {
[…]
