+++ R/sw_combine.R
@@ -16,10 +16,10 @@
 #'   sapwood rings are listed (should be `numeric`).
 #' @param waneyedge Name of the column in `x` indicating the presence
 #'   (`TRUE`)/absence (`FALSE`) of waney edge (should be a `logical` vector).
-#' @param sw_data The name of the sapwood data set to use for modelling. Should
-#'   be one of [sw_data_overview()], or the path to a .csv file with columns
-#'   ´n_sapwood´ and ´count´.
-#' @param credMass A `scalar [0, 1]` specifying the mass within the credible
+#' @param sw_data The name of the sapwood data set to use for modelling. It
+#'   should be one of the data sets listed in [sw_data_overview()], or the name
+#'   of a `data.frame` with sapwood data in columns `n_sapwood` and `count`.
+#' @param cred_mass A `scalar [0, 1]` specifying the mass within the credible
 #'   interval (default = .954).
 #' @param hdi A `logical` parameter. If `TRUE`, the lower and upper limit of the
 #'   highest density interval (credible interval) is given for the combined
@@ -30,12 +30,10 @@
 #'   `FALSE`, a list with numeric output of the modelling process is returned.
 #' @param densfun Name of the density function fitted to the sapwood data set.
 #'   Should be one of:
-#'   * "lognormal" (the default value),
-#'   * "normal",
-#'   * "weibull",
-#'   * "gammma".
-#' @param sep Should be "," (comma)  or ";" (semi-colon) and is used when a
-#'   sapwood data set is provided from user-defined .csv-file.
+#'   * _lognormal_ (the default value),
+#'   * _normal_,
+#'   * _weibull_,
+#'   * _gammma_.
 #'
 #' @return Depends on the value of plot. If `TRUE` a ggplot-style is returned,
 #'   when `FALSE` a `data.frame` with the combined probability and modelling
@@ -47,8 +45,8 @@
 #' @examples
 #' # a data set in which all series have preserved sapwood
 #'
-#' trs_example1
-#' sw_combine(trs_example1, plot = FALSE)
+# trs_example1
+# sw_combine(trs_example1, plot = FALSE)
 #' sw_combine(trs_example1, plot = TRUE)
 #'
 #' # a data set in which one series has an exact felling date (= waney edge preserved)
@@ -59,7 +57,7 @@
 #' # a data set in which multiples series have an exact felling date
 #'
 #' trs_example3
-#' sw_combine(trs_example3, plot= FALSE)
+#' sw_combine(trs_example3, plot = FALSE)
 #'
 sw_combine <- function(x,
                        series = "series",
@@ -68,370 +66,356 @@ sw_combine <- function(x,
                        waneyedge = "waneyedge",
                        sw_data = "Hollstein_1980",
                        densfun = "lognormal",
-                       sep = ";",
-                       credMass = 0.954,
+                       cred_mass = 0.954,
                        hdi = TRUE,
                        plot = FALSE) {
-        x <- as.data.frame(x)
-        n <- nrow(x)
-        endDate <- x[, last]
+  check_input(
+    x = x,
+    series = series,
+    last = last,
+    n_sapwood = n_sapwood,
+    waneyedge = waneyedge,
+    sw_data = sw_data,
+    cred_mass = cred_mass,
+    densfun = densfun
+  )
 
-        if (any(is.na(endDate)) | !is.numeric(endDate)) {
-                stop(
-                        "--> Please check the column with 'end dates'.
-Some values are possibly missing or the values are not numeric"
-                )
-        }
-        swr <- x[, n_sapwood]
[…]
