### 2024-03-19 99ae6ee
defensive programming, snake_case and code stuyle for sw_sum and sw_sum_plot
