+++ R/sw_sum.R
@@ -3,32 +3,32 @@
 #' @param x A `data.frame` containing columns c("series", "swr", "waneyedge" ,
 #'   "last"), Column "sw_model" is optional. Could be the output of read_fh(x,
 #'   header = TRUE)
-#' @param series Name of the column in `x` containing the ID's of the series.
-#' @param last Name of the column in `x`containing the calendar year assigned to
-#'   the last measured ring (should be `numeric`).
+#' @param series Name of the column in `x` where ID's of the tree-ring series
+#'   are listed as `character` values.
+#' @param last Name of the column in `x` where calendar years assigned to the
+#'   last measured ring are listed (should be `numeric`).
 #' @param n_sapwood Name of the column in `x` where the number of observed
-#'   sapwood rings are listed (should be `numeric`).
+#'   sapwood rings are listed (should be `numeric`)
 #' @param waneyedge Name of the column in `x` indicating the presence
-#'   (TRUE)/absence (FALSE) of waney edge (should be a `logical` vector).
-#' @param credMass A `scalar [0, 1]` specifying the mass within the credible
+#'   (`TRUE`)/absence (`FALSE`) of waney edge (should be a `logical`).
+#' @param cred_mass A `scalar [0, 1]` specifying the mass within the credible
 #'   interval (default = .954).
-#' @param sw_data The name of the sapwood data set to use for modelling. Should
-#'   be one of [sw_data_overview()], or the path to a .csv file with columns
-#'   ´n_sapwood´ and ´count´.
+#' @param sw_data The name of the sapwood data set to use for modelling. It
+#'   should be one of the data sets listed in [sw_data_overview()], or the name
+#'   of a `data.frame` with sapwood data in columns `n_sapwood` and `count`.
 #' @param densfun Name of the density function fitted to the sapwood data set.
 #'   Should be one of:
-#'   * "lognormal" (the default value),
-#'   * "normal",
-#'   * "weibull",
-#'   * "gammma".
+#'   * _lognormal_ (the default value),
+#'   * _normal_,
+#'   * _weibull_,
+#'   * _gammma_.
 #' @param plot A `logical`.
 #'   * If `TRUE`, [sw_sum_plot()] is triggered and a ggplot-style graph is
 #'   returned with the summed probability density (SPD).
-#'   * If `FALSE`, a list with the numeric output of the modelling process is returned.
+#'   * If `FALSE`, a list with the numeric output of the modelling process is
+#'   returned.
 #' @param scale_p A `logical`. If `TRUE` the summed probability density is
-#'   scaled to 1.
-#' @param sep Should be "," (comma)  or ";" (semi-colon) and is used when a
-#'   sapwood data set is provided from user-defined .csv-file.
+#'   scaled to 1 (default).
 #'
 #' @description Computes the summed probability density (SPD) for a set of
 #'   felling date ranges.
@@ -41,190 +41,180 @@
 #'
 #' sw_sum(trs_example7, plot = TRUE)
 #'
-sw_sum <- function (x,
-                    series = "series",
-                    last = "last",
-                    n_sapwood = "n_sapwood",
-                    waneyedge = "waneyedge",
-                    sw_data = "Hollstein_1980",
-                    densfun = "lognormal",
-                    credMass = 0.954,
-                    plot = FALSE,
-                    scale_p = FALSE,
-                    sep = ";") {
-        df <- as.data.frame(x)
+sw_sum <- function(x,
+                   series = "series",
+                   last = "last",
[…]
