+++ R/sw_sum_plot.r
@@ -26,77 +26,90 @@ sw_sum_plot <- function(x,
                         dot_col = "steelblue4",
                         dot_size = 2,
                         dot_shape = 21) {
+  # to avoid notes in CMD check
+  spline <- spd_wk <- year <- y <- spd <- spd_mov_av <- NULL
 
-        # to avoid notes in CMD check
-        spline <- SPD_wk <- year <- y <- SPD <- SPD_movAv <- NULL
 
+  pdf_matrix <- as.data.frame(x)
 
-pdf_matrix <- as.data.frame(x)
+  smooth <-
+    as.data.frame(
+      mov_av(pdf_matrix$spd, 11, align = "center", edges = "fill")
+    )
+  names(smooth) <- "spd_mov_av"
 
-smooth <-
-        as.data.frame(
-                movAv(pdf_matrix$SPD, 11, align = "center", edges = "fill")
-                )
-names(smooth) <- "SPD_movAv"
+  pdf_matrix <- cbind(pdf_matrix, smooth)
 
-pdf_matrix <- cbind(pdf_matrix, smooth)
+  spline_int <-
+    as.data.frame(
+      spline(pdf_matrix$year, pdf_matrix$spd_mov_av, method = "natural")
+    )
 
-spline_int <-
-        as.data.frame(
-                spline(pdf_matrix$year, pdf_matrix$SPD_movAv, method = "natural")
-                )
+  plot_range <- range(pdf_matrix$year)
 
-plot_range <- range(pdf_matrix$year)
+  p_max <- max(pdf_matrix$spd)
 
-p_max <- max(pdf_matrix$SPD)
+  fd <- pdf_matrix[, c("year", "spd_wk")]
+  fd <- subset(fd, spd_wk > 0)
+  if (nrow(fd) > 0) {
+    fd <- rep(fd$year, fd$spd_wk)
+    fd <- data.frame(fd)
+    names(fd) <- "year"
+    fd$y <- 1
+    fd <- fd |>
+      dplyr::group_by(year) |>
+      dplyr::mutate(y = cumsum(y)) |>
+      dplyr::left_join(pdf_matrix |> dplyr::select("year", "spd"), by = "year")
+  }
 
-fd <- pdf_matrix[, c("year", "SPD_wk")]
-fd <- subset(fd, SPD_wk>0)
-if (nrow(fd) > 0){
-        fd <- rep(fd$year, fd$SPD_wk)
-        fd <- data.frame(fd)
-        names(fd) <- "year"
-        fd$y <- 1
-        fd <- fd |>
-        dplyr::group_by(year) |>
-        dplyr::mutate(y = cumsum(y)) |>
-        dplyr::left_join(pdf_matrix |> dplyr::select("year", "SPD"), by = "year")
-}
-
-p <-
-     pdf_matrix |>
-        dplyr::select(year, SPD, SPD_movAv) |>
-        ggplot2::ggplot() +
-        ggplot2::geom_col(ggplot2::aes(x = year, y = SPD),
-                          fill = bar_col,
-                          alpha = 0.8,
-                          na.rm = TRUE) +
-        ggplot2::geom_line(data = spline_int,
-                           ggplot2::aes(x = x, y = y),
-                           color = spline_col,
-                           linewidth = 1.5,
-                           alpha = .8,
-                           na.rm = TRUE) +
-        { if (nrow(fd) > 0) ggplot2::geom_point(data = fd,
-                            ggplot2::aes(x = year, y = SPD + y * p_max/20),
-                            size = dot_size,
-                            fill = dot_col,
-                            color = dot_col,
-                            shape = dot_shape,
-                            na.rm = TRUE) } +
-        ggplot2::scale_x_continuous(
[…]
