+++ NAMESPACE
@@ -13,5 +13,6 @@ export(sw_data_overview)
 export(sw_interval)
 export(sw_interval_plot)
 export(sw_model)
+export(sw_model_plot)
 export(sw_sum)
 export(sw_sum_plot)
