### 2024-03-19 c050812
test for sw_model_plot
