### 2024-03-19 4ac771a
implement seq_len() and replace is.na = T by TRUE
