+++ DESCRIPTION
@@ -8,7 +8,11 @@ Authors@R: c(
     person("Koen", "Van Daele", role = "ctb",
         comment = c(ORCID = "0000-0002-8153-2978")),
     person("Ronald", "Visser", role = "ctb",
-        comment = c(ORCID = "0000-0001-6966-1729"))    
+        comment = c(ORCID = "0000-0001-6966-1729")),
+    person("Antonio Jesus", "Pérez-Luque", role = "rev",
+        comment = c(ORCID = "0000-0002-1747-0469")),   
+    person("Nicholas John", "Tierney", role = "rev",
+        comment = c(ORCID = "0000-0003-1460-8722"))   
     )
 Description: fellingdateR is an R package that aims to facilitate the 
   analysis and interpretation of tree-ring data from wooden 
