+++ review_rOpenSci.Rmd
@@ -51,6 +51,12 @@ editor_options:
 > to install the development version of the pkg? I see that in the
 > vignette he indicated the two ways (pak and devtools).
 >
+> ***I work fro a government agency and mostly behind a firewall.
+> Installing packages with devtool::install_packages() often fails
+> (timeout) but works fine with pak::pak(). I have no idea why and it is
+> probably realted to the settings of the firewall, but therefore I
+> prefer pak over devtools now.***
+>
 > -   [ ] **Vignette(s):** demonstrating major functionality that runs
 >     successfully locally
 >
@@ -74,6 +80,8 @@ editor_options:
 > ```
 >
 > Please consider change the three `:::` by `::`
+
+> ***`:::` has been removed from the examples in the vignette***
 >
 > -   [x] **Function Documentation:** for all exported functions
 > -   [x] **Examples:** (that run successfully locally) for all exported
@@ -85,6 +93,8 @@ editor_options:
 >
 > No community guidelines found, only in DESCRIPTION there is a
 > BugReports linked.
+>
+> ***Guidelines have been added to th README***
 >
 > #### Functionality
 >
@@ -106,8 +116,8 @@ editor_options:
 >     -   Please consider to use lowercase name as indicated by 1.1.1
 >         point of the packaging guidelines
 >
->         ***snake_case and lowercase naming has been implemented
->         now.***
+>         ***snake_case and lowercase naming have been implemented
+>         consistently .***
 >
 >     -   Not all the functions follow the object_verb() naming scheme.
 >         For instance, the functions read_fh and get_header. I would
@@ -128,13 +138,15 @@ editor_options:
 >         change cat() or print() for message or warning, as packaging
 >         guidelines indicated.
 >
->         *cat() no longer used in any of the functions*
+>         ***cat() no longer used in any of the functions***
 >
 > Estimated hours spent reviewing: 13
 >
 > -   [x] Should the author(s) deem it appropriate, I agree to be
 >     acknowledged as a package reviewer ("rev" role) in the package
->     DESCRIPTION file.
+>     DESCRIPTION file***.***
+>
+> ***I've included you as a reviewer in the DESCRIPTON***
 >
 > ### Review Comments
 >
@@ -163,6 +175,14 @@ editor_options:
 > -   The data description lacks a consistent pattern. The author
 >     occasionally employs country abbreviations while at other times
 >     refrains from doing so. Let's strive for coherence in this regard.
+>
+>     ***This is because some of de provided sapwood datasets do not
+>     cover a specific country, but rather a region, or are composed of
+>     observations of timber from e.g. Norway but the timbers themselves
+>     have been found in the Netherlands (van Daalen_Norway).*** ***So I
+>     named the sapwood datasets according to their authors (and
+>     publication date if applicable, and a few letters to describe the
+>     region/country they cover.***
 >
 > -   Some code and documentation overpass the 80 characters
 >     (particularly data.R). Please consider adjust to this recommended
@@ -198,6 +218,9 @@ editor_options:
 >     like the rest of the functions in the package (eg. sw_interval and
 >     sw_interval_plot)? This could lighten the function size and also
 >     would be coherent with the rest of the package.
+>
+>     ***There is now a new function named sw_model_plot() as
+>     suggested.***
 >
 > -   The code style isn't consistent. Please consider use the styler
 >     package.
@@ -252,6 +275,8 @@ editor_options:
 >
 > Please consider review the test coverage of the movAv.R function
 > (\~57.14%)
+>
+> ***test coverage increased for mov_av()***
 >
 > ### Goodpractices
 >
@@ -310,6 +335,8 @@ editor_options:
 > R/cor_table.R:195:20
 > ```
 >
[…]
