+++ man/sw_combine_plot.Rd
@@ -4,13 +4,14 @@
 \alias{sw_combine_plot}
 \title{Plot the output of \code{\link[=sw_combine]{sw_combine()}}}
 \usage{
-sw_combine_plot(x, ...)
+sw_combine_plot(x)
 }
 \arguments{
 \item{x}{Output of \code{\link[=sw_combine]{sw_combine()}}.}
 }
 \value{
-A ggplot-style graph.
+A ggplot-style graph, with calendar years on the X-axis and the
+probability (p) on the Y-axis. Dots represent exact felling dates.
 }
 \description{
 This function plots the result of \code{sw_combine()} and the
@@ -19,7 +20,7 @@ with agreement indices for the global model.
 }
 \examples{
 trs_example2
-combo <- sw_combine(trs_example2)
-sw_combine_plot(combo)
+tmp <- sw_combine(trs_example2)
+sw_combine_plot(tmp)
 
 }
