+++ man/sw_sum_plot.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/sw_sum_plot.R
+% Please edit documentation in R/sw_sum_plot.r
 \name{sw_sum_plot}
 \alias{sw_sum_plot}
 \title{Plot the output of \code{\link[=sw_sum]{sw_sum()}}}
@@ -7,10 +7,11 @@
 sw_sum_plot(
   x,
   bar_col = "steelblue",
-  spline_col = "red3",
+  trend_col = "red3",
   dot_col = "steelblue4",
   dot_size = 2,
-  dot_shape = 21
+  dot_shape = 21,
+  window_smooth = 11
 )
 }
 \arguments{
@@ -18,16 +19,21 @@ sw_sum_plot(
 
 \item{bar_col}{The fill color for the bars.}
 
-\item{spline_col}{The line color of the fitted smoothing spline.}
+\item{trend_col}{The line color of the fitted smoothing spline.}
 
 \item{dot_col}{The color of the shapes that represent exact felling dates.}
 
-\item{dot_size}{A size argument for the shapes that represent exact felling dates.}
+\item{dot_size}{A size argument for the shapes that represent exact felling
+dates.}
 
-\item{dot_shape}{Number corresponding to the point symbol available in R for the shapes that represent exact felling dates.}
+\item{dot_shape}{Number corresponding to the point symbol available in R for
+the shapes that represent exact felling dates.}
+
+\item{window_smooth}{A numeric value for the window width of the trend line}
 }
 \value{
-A ggplot style graph.
+A ggplot style graph, with calendar years on the X-axis and the
+probability (p) on the Y-axis.
 }
 \description{
 This function plots the result of \code{sw_sum()} - both the SPD and
