+++ review_rOpenSci.Rmd
@@ -9,6 +9,8 @@ editor_options:
 
 # Reply to review
 
+# Review by Antonio Pérez-Luque
+
 > Sorry for the slight delay in the review. Here it is.
 >
 > ## Package Review
@@ -521,7 +523,7 @@ editor_options:
 >
 >     ***You are probably right. I might have overcomplicated the
 >     sw-functions with this. So I removed this option from this and all
->     other sw_functions. Now a user-defined custom dataset  -
+>     other sw_functions. Now a user-defined custom dataset -
 >     a`data.frame`*** ***with columns `n_sapwood` and `count` - can be
 >     supplied to the function.***
 >
@@ -970,33 +972,42 @@ this is not an option that would be used frequently., and is certainly
 not required. Removing it from the function allows to shorten the code a
 bit.***
 
-# Reviewer 1
+# Review by Nicholas Tierney
 
-Package Review Please check off boxes as applicable, and elaborate in
-comments below. Your review is not limited to these topics, as described
-in the reviewer guide
+```         
+          ## Package Review
+```
 
-Briefly describe any working relationship you have (had) with the
-package authors. ☒ As the reviewer I confirm that there are no conflicts
-of interest for me to review this work (if you are unsure whether you
-are in conflict, please speak to your editor before starting your
-review). Documentation The package includes all the following forms of
-documentation:
+*Please check off boxes as applicable, and elaborate in comments below.
+Your review is not limited to these topics, as described in the reviewer
+guide*
 
-☐ A statement of need: clearly stating problems the software is designed
-to solve and its target audience in README The opening paragraphs of the
-README are good, and I think that this R package solves a challenging
-problem, so firstly, well done! I think could be made a little bit
-clearer in terms of the problem it solves, and the input it takes. While
-I find the photos useful, it initially made me think that this software
-takes images as input. I would suggest something more like what is in
-the vignette to start:
+-   **Briefly describe any working relationship you have (had) with the
+    package authors.**
+-   [x] As the reviewer I confirm that there are no [conflicts of
+    interest](https://devguide.ropensci.org/policies.html#coi) for me to
+    review this work (if you are unsure whether you are in conflict,
+    please speak to your editor *before* starting your review).
 
-fellingdateR offers a set of functions that assist in inferring felling
-date estimates from dated tree-ring series.
+#### Documentation
+
+The package includes all the following forms of documentation:
+
+-   [ ] **A statement of need:** clearly stating problems the software
+    is designed to solve and its target audience in README
+
+The opening paragraphs of the README are good, and I think that this R
+package solves a challenging problem, so firstly, well done! I think
+could be made a little bit clearer in terms of the problem it solves,
+and the input it takes. While I find the photos useful, it initially
+made me think that this software takes images as input. I would suggest
+something more like what is in the vignette to start:
+
+> `fellingdateR` offers a set of functions that assist in inferring
+> felling date estimates from dated tree-ring series.
 
 Then, describe the problem you want to solve, which I think is
-estimating when the timber was cut down. Then show the data, explain
+estimating *when* the timber was cut down. Then show the data, explain
 what the columns mean, and how this might be a typical example of dated
 tree-ring series data.
 
@@ -1005,12 +1016,12 @@ problem the package solves.
 
 The rest of the first paragraph:
 
-The presence of (partially) preserved sapwood or waney edge allows to
-estimate a range for the actual felling date, for individual series as
-well as for a group of timbers. Furthermore, an additional function
-provides a tool to sum sapwood probability distributions, comparable to
-'summed probability densities' commonly applied to sets of radiocarbon
-(14C) dates.
[…]
