+++ review_rOpenSci.Rmd
@@ -1,1708 +1,0 @@
----
-title: "review_rOpenSci"
-output: html_document
-date: "2024-02-24"
-editor_options: 
-  markdown: 
-    wrap: 72
----
-
-# Reply to review
-
-# Review by Antonio Pérez-Luque
-
-> Sorry for the slight delay in the review. Here it is.
->
-> ## Package Review
->
-> *Please check off boxes as applicable, and elaborate in comments
-> below. Your review is not limited to these topics, as described in the
-> reviewer guide*
->
-> -   **Briefly describe any working relationship you have (had) with
->     the package authors.**
-> -   [x] As the reviewer I confirm that there are no [conflicts of
->     interest](https://devguide.ropensci.org/policies.html#coi) for me
->     to review this work (if you are unsure whether you are in
->     conflict, please speak to your editor *before* starting your
->     review).
->
-> #### Documentation
->
-> The package includes all the following forms of documentation:
->
-> -   [ ] **A statement of need:** clearly stating problems the software
->     is designed to solve and its target audience in README
->
-> I think the target audience of the pkg would be indicated more
-> explicitly.
->
-> The images are not very illustrative of what the package does, they
-> simply seem to indicate that they are wooden tokens. Perhaps it would
-> be interesting to think of a schematic or similar, but the photos do
-> not seem to provide much.
->
-> -   [ ] **Installation instructions:** for the development version of
->     package and any non-standard dependencies in README
->
-> In the installation instructions, it might be more effective to
-> provide a link to the GitHub repository of the package rather than
-> displaying the raw GitHub link itself.
->
-> Why does the author prefer to use the package pak rather than devtools
-> to install the development version of the pkg? I see that in the
-> vignette he indicated the two ways (pak and devtools).
->
-> ***I work fro a government agency and mostly behind a firewall.
-> Installing packages with devtool::install_packages() often fails
-> (timeout) but works fine with pak::pak(). I have no idea why and it is
-> probably realted to the settings of the firewall, but therefore I
-> prefer pak over devtools now.***
->
-> -   [ ] **Vignette(s):** demonstrating major functionality that runs
->     successfully locally
->
-> The vignette outlines the primary functions but lacks a
-> straightforward workflow for beginners.
->
-> There are missing citations in the vignette. Please add a reference
-> list at the end.
->
-> In the vignette of the website it appears:
->
-> ```         
-> sw_sum(fellingdateR:::trs_example7, plot = TRUE)
-> ```
->
-> but when I run locally i got:
->
-> ```         
-> sw_sum(fellingdateR:::trs_example7, plot = TRUE)
-> Error in as.data.frame(x) : object 'trs_example7' not found
-> ```
->
-> Please consider change the three `:::` by `::`
-
-> ***`:::` has been removed from the examples in the vignette***
->
-> -   [x] **Function Documentation:** for all exported functions
-> -   [x] **Examples:** (that run successfully locally) for all exported
->     functions
-> -   [ ] **Community guidelines:** including contribution guidelines in
->     the README or CONTRIBUTING, and DESCRIPTION with `URL`,
->     `BugReports` and `Maintainer` (which may be autogenerated via
->     `Authors@R`).
->
-> No community guidelines found, only in DESCRIPTION there is a
-> BugReports linked.
->
-> ***Guidelines have been added to th README***
->
-> #### Functionality
->
-> -   [x] **Installation:** Installation succeeds as documented.
-> -   [x] **Functionality:** Any functional claims of the software have
->     been confirmed.
-> -   [x] **Performance:** Any performance claims of the software have
->     been confirmed.
-> -   [x] **Automated tests:** Unit tests cover essential functions of
->     the package and a reasonable range of inputs and conditions. All
[…]
