### 2024-03-19 04052bd
updating documentation for fucntions, datasets and vignette
