+++ man/Hollstein_1980.Rd
@@ -27,7 +27,7 @@ Sapwood data set for South and Central Germany published by Hollstein in 1980.
 \examples{
 sw_data_info("Hollstein_1980")
 
-sw_model("Hollstein_1980")
+sw_model("Hollstein_1980", plot = TRUE)
 
 }
 \keyword{Hollstein_1980}
