### 2024-03-19 54655bc
implementing some defensve programming
