+++ DESCRIPTION
@@ -28,7 +28,7 @@ Language: en-GB
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
 Depends: 
-    R (>= 2.10)
+    R (>= 4.1.0)
 LazyData: true
 Imports: 
     dplyr,
