+++ R/cor_table.R
@@ -83,7 +83,8 @@
 #' Doel1_trs <- read_fh(Doel1, verbose = FALSE)
 #' # crossdating ring-width series from Doel 1 against each other:
 #'
-#' cor_results <- cor_table(Doel1_trs, output = "table", min_overlap = 80)
+#' cor_results <- cor_table(Doel1_trs, output = "table", min_overlap = 80,
+#' sort_by = "t_Ho", remove_duplicates = TRUE)
 #'
 cor_table <-
   function(x,
