+++ R/fh_header.R
@@ -5,7 +5,7 @@
 #'   the `read_fh()` function, which stores the HEADER fields from the .fh file
 #'   as attributes of the `data.frame` with the measurement data it returns.
 #'
-#' @param rwl The output of `read_fh(x, header = TRUE)`, a `data.frame` of class
+#' @param x The output of `read_fh(x, header = TRUE)`, a `data.frame` of class
 #'   `rwl`.
 #'
 #' @return A `data.frame` with 29 header fields.
@@ -17,18 +17,18 @@
 #' @export
 
 fh_header <-
-  function(rwl) {
-    if (!is.data.frame(rwl) || !inherits(rwl, "rwl")) {
+  function(x) {
+    if (!is.data.frame(x) || !inherits(x, "rwl")) {
       stop("Input should be a data.frame of class 'rwl'")
     }
 
-    attr(rwl, "row.names") <- NULL
-    attr(rwl, "po") <- NULL
-    attr(rwl, "class") <- NULL
-    attr(rwl, "names") <- NULL
+    attr(x, "row.names") <- NULL
+    attr(x, "po") <- NULL
+    attr(x, "class") <- NULL
+    attr(x, "names") <- NULL
 
-    tmp <- attributes(rwl)
-    tmp <- data.frame(tmp)
+    out <- attributes(x)
+    out <- data.frame(out)
 
-    return(tmp)
+    return(out)
   }
