+++ man/sw_model_plot.Rd
@@ -34,6 +34,9 @@ tmp <- sw_model(
   cred_mass = .95,
   plot = FALSE
 )
-sw_model_plot(tmp)
+sw_model_plot(tmp,
+              bar_fill = "forestgreen",
+              bar_color = "darkgreen",
+              line_color = "tomato3")
 
 }
