+++ man/sw_sum.Rd
@@ -71,7 +71,7 @@ felling date ranges.
 \examples{
 trs_example7
 
-sw_sum(trs_example7, plot = TRUE)
+sw_sum(trs_example7, densfun = "lognormal", cred_mass = 0.63, plot = FALSE)
 
 }
 \seealso{
