+++ R/sw_interval_plot.R
@@ -120,14 +120,7 @@ sw_interval_plot <-
       hjust = 1,
       vjust = 1
     ) +
-    ggplot2::theme(
-      axis.text = ggplot2::element_text(size = 10),
-      panel.grid.minor.y = ggplot2::element_blank(),
-      panel.grid.major.y = ggplot2::element_blank(),
-      legend.position = "none",
-      strip.background = ggplot2::element_blank(),
-      strip.text.y = ggplot2::element_blank()
-    ) +
+    theme_fdr() +
     ggplot2::ylab("p") +
     ggplot2::xlab("calendar year")
 
