+++ NEWS.md
@@ -8,3 +8,10 @@ fellingdateR 0.0.0.9003 (2023-11-20)
   * First draft op paper for Journal of Open Software in /paper/paper.md
   
   
+fellingdateR 0.0.0.9004 (2024-03-20)
+====================================
+### REVIEW
+  * new version of the package after review
+  * review history: https://github.com/ropensci/software-review/issues/618
+  
+ 
