+++ README.md
@@ -13,16 +13,16 @@ coverage](https://codecov.io/gh/hanecakr/fellingDateR/branch/main/graph/badge.sv
 Review](https://badges.ropensci.org/618_status.svg)](https://github.com/ropensci/software-review/issues/618)
 <!-- badges: end -->
 
-This R-package offers a set of functions that can assist you to infer
+This package offers a set of functions that can assist you to infer
 felling date estimates from dated tree-ring series. The presence of
-(partially) preserved sapwood or waney edge allows to estimate a range
-for the actual felling date, for individual series as well as for a
-group of timbers. Furthermore, an additional function provides a tool to
-sum sapwood probability distributions, comparable to ‘summed probability
-densities’ commonly applied to sets of radiocarbon (<sup>14</sup>C)
-dates.
+(partially) preserved sapwood allows to estimate the missing number of
+sapwood rings ( ? in figure below), and to report an interval in which
+the actual felling date is situated.
 
-![](man/figures/roof_cores.png)
+![](man/figures/cross-section.png)
+
+This procedure can be implemented for individual series as well as for a
+group of timbers.
 
 Where it can be assumed that a group of historical timbers were all
 felled at the same time (i.e. in the same year), but due to the absence
@@ -34,6 +34,10 @@ Taking into account the observed number of sapwood rings on all samples
 and combining them into a single estimate, is likely to provide a more
 accurate and precise estimate of the felling date year for the group of
 timbers under study.
+
+Furthermore, an additional function provides a tool to sum sapwood
+probability distributions, comparable to ‘summed probability densities’
+commonly applied to sets of radiocarbon (<sup>14</sup>C) dates.
 
 ## Installation
 
@@ -54,7 +58,6 @@ five dated tree-ring series is computed:
 library(fellingdateR)
 
 ## a data set where all series have partially preserved sapwood:
-load("R/sysdata.rda")
 trs_example1
 #>   series last n_sapwood waneyedge
 #> 1 trs_01 1000         5     FALSE
@@ -79,9 +82,131 @@ range, was published by Hollstein in 1980:
 
 ``` r
 sw_model("Hollstein_1980")
+#> $sapwood_data
+#> [1] "Hollstein_1980"
+#> 
+#> $n
+#> [1] 490
+#> 
+#> $range
+#>   min  mean   max 
+#>  6.00 25.05 67.00 
+#> 
+#> $density_function
+#> [1] "lognormal"
+#> 
+#> $fit_parameters
+#>     meanlog       sdlog   
+#>   2.81357890   0.41620736 
+#>  (0.01880233) (0.01329526)
+#> 
+#> $sapwood_model
+#>     n_sapwood            p    model_fit count
+#> 1           1 1.143856e-10 5.604897e-08    NA
+#> 2           2 1.107667e-06 5.427569e-04    NA
+#> 3           3 6.572161e-05 3.220359e-02    NA
+#> 4           4 6.697648e-04 3.281848e-01    NA
+#> 5           5 2.917879e-03 1.429761e+00    NA
+#> 6           6 7.845478e-03 3.844284e+00     4
+#> 7           7 1.558774e-02 7.637994e+00     9
+#> 8           8 2.528777e-02 1.239101e+01    18
+#> 9           9 3.557549e-02 1.743199e+01    21
+#> 10         10 4.511128e-02 2.210453e+01    21
+#> 11         11 5.291838e-02 2.593001e+01    21
[…]
