### 2024-03-19 1e569e4
edit README (include plot = TRUE)
