+++ README.md
@@ -81,132 +81,10 @@ The sapwood data used in the example above to estimate the felling date
 range, was published by Hollstein in 1980:
 
 ``` r
-sw_model("Hollstein_1980")
-#> $sapwood_data
-#> [1] "Hollstein_1980"
-#> 
-#> $n
-#> [1] 490
-#> 
-#> $range
-#>   min  mean   max 
-#>  6.00 25.05 67.00 
-#> 
-#> $density_function
-#> [1] "lognormal"
-#> 
-#> $fit_parameters
-#>     meanlog       sdlog   
-#>   2.81357890   0.41620736 
-#>  (0.01880233) (0.01329526)
-#> 
-#> $sapwood_model
-#>     n_sapwood            p    model_fit count
-#> 1           1 1.143856e-10 5.604897e-08    NA
-#> 2           2 1.107667e-06 5.427569e-04    NA
-#> 3           3 6.572161e-05 3.220359e-02    NA
-#> 4           4 6.697648e-04 3.281848e-01    NA
-#> 5           5 2.917879e-03 1.429761e+00    NA
-#> 6           6 7.845478e-03 3.844284e+00     4
-#> 7           7 1.558774e-02 7.637994e+00     9
-#> 8           8 2.528777e-02 1.239101e+01    18
-#> 9           9 3.557549e-02 1.743199e+01    21
-#> 10         10 4.511128e-02 2.210453e+01    21
-#> 11         11 5.291838e-02 2.593001e+01    21
-#> 12         12 5.847980e-02 2.865510e+01    27
-#> 13         13 6.168335e-02 3.022484e+01    27
-#> 14         14 6.270361e-02 3.072477e+01    37
-#> 15         15 6.187987e-02 3.032113e+01    25
-#> 16         16 5.961755e-02 2.921260e+01    21
-#> 17         17 5.632071e-02 2.759715e+01    24
-#> 18         18 5.235228e-02 2.565261e+01    29
-#> 19         19 4.801543e-02 2.352756e+01    23
-#> 20         20 4.354902e-02 2.133902e+01    18
-#> 21         21 3.913128e-02 1.917433e+01    19
-#> 22         22 3.488778e-02 1.709501e+01    26
-#> 23         23 3.090081e-02 1.514139e+01    17
-#> 24         24 2.721872e-02 1.333717e+01    19
-#> 25         25 2.386428e-02 1.169350e+01    14
-#> 26         26 2.084176e-02 1.021246e+01     9
-#> 27         27 1.814260e-02 8.889875e+00    11
-#> 28         28 1.574987e-02 7.717438e+00     8
-#> 29         29 1.364165e-02 6.684408e+00     6
-#> 30         30 1.179345e-02 5.778789e+00     3
-#> 31         31 1.017998e-02 4.988188e+00     3
-#> 32         32 8.776330e-03 4.300402e+00     3
-#> 33         33 7.558758e-03 3.703792e+00     5
-#> 34         34 6.505127e-03 3.187512e+00    NA
-#> 35         35 5.595166e-03 2.741631e+00     4
-#> 36         36 4.810558e-03 2.357174e+00     5
-#> 37         37 4.134922e-03 2.026112e+00     4
-#> 38         38 3.553733e-03 1.741329e+00     1
-#> 39         39 3.054196e-03 1.496556e+00     3
-#> 40         40 2.625106e-03 1.286302e+00    NA
-#> 41         41 2.256692e-03 1.105779e+00     1
-#> 42         42 1.940470e-03 9.508301e-01     1
-#> 43         43 1.669088e-03 8.178532e-01    NA
[…]
