### 2024-03-20 4f5cef1
implementing input checks with helper functions For sw_model & sw_interval
