+++ R/cor_table.R
@@ -85,6 +85,7 @@
 #'
 #' cor_results <- cor_table(Doel1_trs, output = "table", min_overlap = 80,
 #' sort_by = "t_Ho", remove_duplicates = TRUE)
+#' head(cor_results,20)
 #'
 cor_table <-
   function(x,
