+++ man/mov_av.Rd
@@ -38,7 +38,9 @@ This vectorized function computes a running mean/moving average
 over a continuous (time) series with a specified window width.
 }
 \examples{
-num_vec <- rnorm(100)
-mov_av(num_vec, w = 20)
+num_vec <- sample(seq(50, 100, 1), 100, replace = TRUE)
+filtered <- mov_av(num_vec, w = 5, align = "center", edges = "nofill")
+plot(num_vec, type = "l")
+lines(filtered, col = "darkblue")
 
 }

[…]
