+++ README.md
@@ -13,36 +13,37 @@ coverage](https://codecov.io/gh/hanecakr/fellingDateR/branch/main/graph/badge.sv
 Review](https://badges.ropensci.org/618_status.svg)](https://github.com/ropensci/software-review/issues/618)
 <!-- badges: end -->
 
-This package offers a set of functions that can assist you to infer
-felling date estimates from dated tree-ring series. The presence of
-(partially) preserved sapwood allows to estimate the missing number of
-sapwood rings ( ? in figure below), and to report an interval in which
-the actual felling date is situated.
+This R-package offers a suite of functions designed to assist
+dendrochronologists in inferring estimates for felling dates, derived
+from dated tree-ring series. The presence of partially preserved sapwood
+(*sw*) allows to estimate the missing number of sapwood rings ( ? in
+figure below), and to report an interval in which the actual felling
+date (*fd*) likely falls.
 
 ![](man/figures/cross-section.png)
 
-This procedure can be implemented for individual series as well as for a
-group of timbers.
+This procedure can be applied to individual series as well as to a
+collection of timbers.
 
-Where it can be assumed that a group of historical timbers were all
-felled at the same time (i.e. in the same year), but due to the absence
-of the bark/cambial zone (waney edge) and the last formed tree ring this
-cannot be assessed, the preserved sapwood rings can be used to infer a
-date range for the felling date.
+In cases where it can be assumed that a group of historical timbers were
+all felled simultaneously (i.e., in the same year), but due to the
+absence of the bark/cambial zone (waney edge) and the final formed tree
+ring, this cannot be determined, the preserved sapwood rings can be used
+to establish a date range for the felling event.
 
-Taking into account the observed number of sapwood rings on all samples
-and combining them into a single estimate, is likely to provide a more
-accurate and precise estimate of the felling date year for the group of
-timbers under study.
+Taking into account the observed number of sapwood rings across all
+analysed samples and combining them into a single estimate, a more
+accurate and precise estimation of the felling date year for the group
+of timbers under study is likely to be obtained.
 
-Furthermore, an additional function provides a tool to sum sapwood
-probability distributions, comparable to ‘summed probability densities’
+An additional function offers a tool for summing sapwood probability
+distributions, comparable to ‘summed probability densities (SPD)’
 commonly applied to sets of radiocarbon (<sup>14</sup>C) dates.
 
 ## Installation
 
-You can install the development version of fellingdateR from
-[GitHub](https://github.com/) with:
+You can install the development version of `fellingdateR` from
+[GitHub](https://github.com/hanecakr/fellingdateR/) with:
 
 ``` r
 #install.packages("pak")
@@ -77,8 +78,9 @@ The light grey distributions represent the probability density function
 of the felling date range for each individual series. The dark grey
 distribution is the combined estimate for a common felling date.
 
-The sapwood data used in the example above to estimate the felling date
-range, was published by Hollstein in 1980:
+The sapwood model used in the example above to estimate the felling date
+range, was published by [Hollstein
+(1980)](https://search.worldcat.org/nl/title/6391864):
 
 ``` r
 sw_model("Hollstein_1980", plot = TRUE)
@@ -93,26 +95,26 @@ datasets** distributed with the package in the
 [documentation](https://hanecakr.github.io/fellingdateR/reference/).
 
 Some practical examples are provided in the
-[`getting started`](https://hanecakr.github.io/fellingdateR/articles/getting_started.html)
+[`Get started`](https://hanecakr.github.io/fellingdateR/articles/getting_started.html)
 vignette.
 
-## Genreal workflow
+## The workflow
 
 The `fellingdateR`-package allows to fully document the methodology to
[…]
