### 2024-03-20 #618 @hanecakr
Dear @ajpelu. You comments and suggestions in your review report have been most helpfull to improve the package. Thank you for your time investment!
I've copy-pasted the review report and inserted my replies below.

> ## Package Review
>
> *Please check off boxes as applicable, and elaborate in comments
> below. Your review is not limited to these topics, as described in the
> reviewer guide*
>
> -   **Briefly describe any working relationship you have (had) with
>     the package authors.**
> -   [x] As the reviewer I confirm that there are no [conflicts of
>     interest](https://devguide.ropensci.org/policies.html#coi) for me
>     to review this work (if you are unsure whether you are in
>     conflict, please speak to your editor *before* starting your
>     review).
>
> #### Documentation
>
> The package includes all the following forms of documentation:
>
> -   [ ] **A statement of need:** clearly stating problems the software
>     is designed to solve and its target audience in README
>
> I think the target audience of the pkg would be indicated more
> explicitly.
>
> The images are not very illustrative of what the package does, they
> simply seem to indicate that they are wooden tokens. Perhaps it would
> be interesting to think of a schematic or similar, but the photos do
> not seem to provide much.
>
> -   [ ] **Installation instructions:** for the development version of
>     package and any non-standard dependencies in README
>
> In the installation instructions, it might be more effective to
> provide a link to the GitHub repository of the package rather than
> displaying the raw GitHub link itself.
>
> Why does the author prefer to use the package pak rather than devtools
> to install the development version of the pkg? I see that in the
> vignette he indicated the two ways (pak and devtools).
>
> ***I serve at a government agency and mostly work from behind a
> firewall. Installing packages with devtool::install_packages() often
> fails (timeout) but miraculously works fine with pak::pak(). I have no
> idea why and it is probably related to the settings of the firewall,
> but therefore I prefer pak over devtools now.***
>
> -   [ ] **Vignette(s):** demonstrating major functionality that runs
>     successfully locally
>
> The vignette outlines the primary functions but lacks a
> straightforward workflow for beginners.
>
> There are missing citations in the vignette. Please add a reference
> list at the end.
>
> ***references added as url's to the publication***
>
> In the vignette of the website it appears:
>
> ```
> sw_sum(fellingdateR:::trs_example7, plot = TRUE)
> ```
>
> but when I run locally i got:
>
> ```
> sw_sum(fellingdateR:::trs_example7, plot = TRUE)
> Error in as.data.frame(x) : object 'trs_example7' not found
> ```
>
> Please consider change the three `:::` by `::`

> ***`redundant use od :::` has been removed from the examples in the
> vignette***
>
> -   [x] **Function Documentation:** for all exported functions
> -   [x] **Examples:** (that run successfully locally) for all exported
>     functions
> -   [ ] **Community guidelines:** including contribution guidelines in
>     the README or CONTRIBUTING, and DESCRIPTION with `URL`,
>     `BugReports` and `Maintainer` (which may be autogenerated via
>     `Authors@R`).
>
> No community guidelines found, only in DESCRIPTION there is a
> BugReports linked.
>
> ***Guidelines have been added to th README***
>
> #### Functionality
>
> -   [x] **Installation:** Installation succeeds as documented.
> -   [x] **Functionality:** Any functional claims of the software have
>     been confirmed.
> -   [x] **Performance:** Any performance claims of the software have
>     been confirmed.
> -   [x] **Automated tests:** Unit tests cover essential functions of
>     the package and a reasonable range of inputs and conditions. All
>     tests pass on the local machine.
>
> All test pass. The coverage is 77.44 % (using
> devtools::test_coverage())
>
> -   [ ] **Packaging guidelines**: The package conforms to the rOpenSci
>     packaging guidelines.
>
>     -   Please consider to use lowercase name as indicated by 1.1.1
>         point of the packaging guidelines
>
>         ***snake_case and lowercase naming have been implemented
>         consistently***
>
>     -   Not all the functions follow the object_verb() naming scheme.
>         For instance, the functions read_fh and get_header. I would
>         suggest to author rename them to fh_read and fh_header
>
>         ***Changed get_header to fh_header, but to highlight the close
>         similarity to the original read.fh() function in the dplR
>         package, I would like to keep read_fh as the function name.***
>
>     -   The movAv function's name deviates from the consistent use of
>         snake_case observed in the rest of the functions. To maintain
>         uniformity, it would be preferable to rename it in accordance
>         with the snake_case as indicated in the packaging guidelines
>
>         ***Changed to mov_av()***
>
>     -   In the get_header, movAv, read_fh functions please consider
>         change cat() or print() for message or warning, as packaging
>         guidelines indicated.
>
>         ***cat() no longer used in any of the functions***
>
> Estimated hours spent reviewing: 13
>
> -   [x] Should the author(s) deem it appropriate, I agree to be
>     acknowledged as a package reviewer ("rev" role) in the package
>     DESCRIPTION file***.***
>
> ***I've included you as a reviewer in the DESCRIPTON***
>
> ### Review Comments
>
> Thank you for your contribution with the fellingdateR package. It
> provides a valuable suite of functions for estimating, reporting, and
> combining felling dates of historical tree-ring series. The
> comprehensive set of tools makes it easier to infer felling date
> estimates, considering factors like preserved sapwood or waney edge.
>
> Given my expertise and interest in dendrochronology, I thoroughly
> enjoyed reviewing this package. The functions enable users to estimate
> felling date ranges for both individual series and groups of timbers,
> offering a valuable feature for summing sapwood probability
> distributions.
>
> This being my first software review for Ropensci, I apologize for any
> potential errors or unclear comments. While I've provided various
> comments (a lot of them), feel free to focus on those that are most
> relevant or useful. I commend you for effectively addressing a
> challenging problem with this package. If you need clarification or
> assistance implementing the suggestions, please reach out. Thank you
> for submitting this software; the review process was a pleasure!
>
> ### General comments
>
> -   The data description lacks a consistent pattern. The author
>     occasionally employs country abbreviations while at other times
>     refrains from doing so. Let's strive for coherence in this regard.
>
>     ***This is because some of de provided sapwood datasets do not
>     cover a specific country, but rather a region, or are composed of
>     observations of timber from e.g. Norway but the timbers themselves
>     have been found in the Netherlands (van Daalen_Norway).*** ***So I
>     named the sapwood datasets according to their authors (and
>     publication date if applicable), and a few letters to describe the
>     region/country they cover or method used.***
>
>     ***The names of the data sets can be easily copied from the output
>     of `sw-data_overview()`***
>
> -   Some code and documentation overpass the 80 characters
>     (particularly data.R). Please consider adjust to this recommended
>     length.
>
>     ***All code has been styled using the `styler`package , and should
>     now adhere to to the Tidyverse styleguide***.
>
> -   The README file does not contain any information about how to
>     cite. I found that one in the CITATION file and also in the
>     website of the package. According to Ropensci packaging guidelines
>     the README need to include information about how to cite.
>
>     ***Citation guidelines have been added tot the README***
>
> -   Please consider generate a theme for all plots of your package in
>     a utils.R function
>
>     ***`theme_fdr()` is now part of `helper-functions.R` and
>     implemented in all plotting functions.***
>
> -   I found very useful the /paper/paper.md, and I think much of this
>     information would improve also the documentation and the website
>     of the package. For instance, the fellingdateR_workflow is very
>     useful to understand how the package works. Would the author
>     consider include it in the REAMDE or in the DESCRIPTION of the
>     package instead of the current pictures? I think the former is
>     more informative than the latter
>
>     ***I've inserted the workflow in the README. The paper.md file is
>     a first draft for a paper to be submitted to JOSS, after software
>     review..***
>
> -   Why is there no auxiliary function to sw_model (eg. sw_model_plot)
>     like the rest of the functions in the package (eg. sw_interval and
>     sw_interval_plot)? This could lighten the function size and also
>     would be coherent with the rest of the package.
>
>     ***There is now a new function named sw_model_plot() as
>     suggested.***
>
> -   The code style isn't consistent. Please consider use the styler
>     package.
>
>     ***I've used the styler package to reformat all code.***
>
> -   I didn't find top-level documentation:
>
>     ***I've added a `fellingdateR-package.R` file with top-level
>     documentation.***
>
> ```
> ?fellingdateR
> #No documentation for ‘fellingdateR’ in specified packages and libraries:
> #you could try ‘??fellingdateR’```
> Pleas consider generate ?fellingdateR or `?fellingdateR-package`. See  https://devguide.ropensci.org/building.html#general
> ```
>
> -   The package has a well-structured website, however I thin much of
>     the information included in the paper.md is not in the website.
>     For instance the background and Statement of need within the
>     paper.md is very useful and I think it would help to the potential
>     users of the pacakge. Please consider include in the README or in
>     an specific vignette.
>
> ***The README has been rewritten, including some parts of the
> paper.md.***
>
> -   The vignette "getting started" includes some citations (eg.
>     Hollstein 1980) but not a full reference for this citations.
>     Please include them at the end of the vignette.
>
>     ***Thanks! Bibliographic references are included as url's/doi's
>     now.***
>
> -   Please be consistent with the use of = or \<-, but avoid mixed
>     them. For instance in hdi.R
>
>     ***Now there is a consistent use of `<-` as an assignment
>     operator***
>
> -   **spell check**: DESCRIPTION does not contain 'Language' field.
>     Defaulting to 'en-US'. Please consider to specify.
>
>     ***en-GB is now set as 'Language'***
>
> -   Is there **code duplication** in the package that should be
>     reduced? Yes for instance the ggthemes within the plot functions
>
>     ***`theme_fdr()` is now part of `helper-functions.R` and
>     implemented in all plotting functions.***
>
> ### Coverage
>
> Please consider review the test coverage of the movAv.R function
> (\~57.14%)
>
> ***test coverage increased for mov_av()***
>
> ### Goodpractices
>
> After run the `goodpractices::gp()` several code lines does not pass
> the check:
>
> ✖ write short and simple functions. These functions have high
> cyclomatic complexity (\>50): read_fh (150). You can make them easier
> to reason about by encapsulating distinct steps of your function into
> subfunctions.
>
> ***This is particularly true for the read_fh function. Since the .fh
> file structure is quite flexible and combines both meta-data and
> measurement data into one format, the code is quite lengthy.
> Furthermore, in the fellingdateR package I build upon the original
> code of the read.fh function from the dplR package. I would prefer to
> stay a close as possible to the original code in the dplR package in
> order to facilitate future cooperation and possible integration of
> both functions.***
>
> ✖ use '\<-' for assignment instead of '='. '\<-' is the standard, and
> R users and developers are used it and it is easier to read your code
> for them if you use '\<-'.
>
> ***Now there is a consistent use of \<- as an assignment operator***
>
> ```
> R/cor_table.R:108:18
> R/cor_table.R:109:22
> R/cor_table.R:112:22
> R/hdi.R:48:12
> R/hdi.R:49:14
> ... and 2 more lines
> ```
>
> ✖ avoid long code lines, it is bad for readability. Also, many people
> prefer editor windows that are about 80 characters wide. Try make your
> lines shorter than 80 characters

> ***styler package was used to reformat all code***
>
> ```
> data-raw/DATASETS.Rmd:17:81
> data-raw/DATASETS.Rmd:19:81
> data-raw/DATASETS.Rmd:35:81
> data-raw/DATASETS.Rmd:51:81
> data-raw/DATASETS.Rmd:68:81
> ... and 153 more lines
> ```
>
> ✖ avoid sapply(), it is not type safe. It might return a vector, or a
> list, depending on the input data. Consider using vapply() instead.
>
> ```
> R/cor_table.R:192:20
> R/cor_table.R:195:20
> ```
>
> ***sapply() no longer used***
>
> ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and
> 1:NCOL(...) expressions. They are error prone and result 1:0 if the
> expression on the right hand side is zero. Use seq_len() or
> seq_along() instead.
>
> ```
> R/fd_report.R:162:19
> R/hdi.R:54:16
> R/sw_combine.R:123:27
> R/sw_combine.R:180:27
> R/sw_combine.R:361:35
> ... and 1 more lines
> ```
>
> ***seq_len() is now used instead of 1:length(...)***
>
> ✖ fix this R CMD check NOTE: Note: found 11 marked UTF-8 strings
>
> ✖ avoid 'T' and 'F', as they are just variables which are set to the
> logicals 'TRUE' and 'FALSE' by default, but are not reserved words and
> hence can be overwritten by the user. Hence, one should always use
> 'TRUE' and 'FALSE' for the logicals.
>
> ***`TRUE/FALSE` now used consistently***
>
> ```
> R/sw_combine_plot.R:NA:NA
> R/sw_combine_plot.R:NA:NA
> R/sw_combine_plot.R:NA:NA
> R/sw_combine_plot.R:NA:NA
> R/sw_combine_plot.R:NA:NA
> ... and 12 more lines
> ```
>
> #### `read_fh.R`
>
> -   Whats new funtionalities include read_fh not covered by
>     dplR::read.fh()?
>
>     ***dplR::read.fh() concentrates on extracting the measurement
>     data. The fellingdateR::read_fh() function extracts also the
>     descriptive (meta-)data from the HEADER fields in a .fh file. This
>     is not possible with the dplR::read.fh function.***
>
>     ***Furthermore the fellingdateR::read_fh function allows to read
>     data in CHRON or HALF-CHRONO format.***
>
>     ***read.fh() also throws errors when header fields include Capital
>     letters (depends on the software used to produce the .fh files:
>     TSAP, PAST, ...). read_fh() is case-insensitive.***
>
> -   There are several code commented that could be removed. eg.
>     `# NEW: verbose = TRUE, header = FALSE` Please revise them
>
>     ***Indeed. The comments highlight what sections have been added
>     compared to the original dplR::read.fh function. I've removed the
>     \`NEW:\` comments in the current version of read_fh()***
>
> -   The nomenclature of the objetcs exhibits a mix of snake_case and
>     dot.case. Let's ensure consistency in the naming convention.
>
>     ***In order to stay as close as possible to the original
>     dplR::read.fh function (and to facilitate possible future
>     integration of both functions) I haven't changed any of the
>     variable names in the code. For all other functions in the package
>     snake_case was implemented.***
>
> -   The function appears to be extensive. Has the author explored the
>     possibility of breaking it down into several auxiliary functions
>     to enhance readability and maintainability?
>
>     ***See reply on bullet point 3***
>
> -   In the function documentation, the author mentions the Heidelberg
>     format file. It would be beneficial to include a link to the
>     specifications of this format file for reference, eg.
>     <https://www.treeringsociety.org/resources/SOM/Brewer_Murphy_SupplementaryMaterial.pdf>
>
>     ***Perfect! I've added the link in the description***
>
> #### `get_header.R`
>
> -   Please consider include some defensive programming such us:
>
> ```
> if (!is.data.frame(rwl) || !inherits(rwl, "rwl")) {
>     stop("Input should be a data.frame of class 'rwl'")
>   }
> ```
>
> -   avoid `cat` as indicated in packaging guidelines
>
>     ***The function now starts with a check of the input and no longer
>     requires cat (or equivalent).***
>
>     ***The get_header function was renamed to `fh_header`.***
>
> #### `cor_table.R`
>
> -   Very interesting and useful function. Has the author contemplated
>     breaking down this function into more modular components? For
>     instance, each of the correlation values could potentially be
>     extracted as auxiliary functions, providing users with the
>     flexibility to call them individually as needed.
>
>     ***Certainly considered, but this would require a full rewrite of
>     the current function. Then it should also include the possibility
>     to use the separate correlation values for crossdating a series on
>     all positions along a reference series. For this package however,
>     the goal is to check the date assigned to the tree-ring series and
>     to find the chronology that matches best in order to decide which
>     sapwood model to use (depending on timber provenance). So I would
>     like to keep the function as it is, and have the breakdown into
>     modular components for each correlation value on the todo-list for
>     a future project/package.***
>
> -   The nomenclature of arguments exhibits a mix of snake_case and
>     dot.case. Let's ensure consistency in the naming convention***.***
>
>     ***now all snake_case***
>
> -   Remove code comments in the argument section, eg***.***
>
>     ***Removed***
>
> ```
> ...
> output = "table",
> #c("matrix", "table")
> ...
> ```
>
> -   In the documentation for the argument `sort_by`, it would be
>     useful to add the choices to avoid errors. For instance,
>
> ```
> #' @param sort_by The correlation value by which the output is sorted for each
> #'   series in `x`. One of "r_pearson", "t_St", "glk", "glk_p", "t_BP", "t_Ho". Default to "t_Ho"
> ```
>
> ***Added to the documentation***
>
> -   Ensure uniformity in assignment by using the same operator
>     consistently---avoid mixing `<-` and `=`
>
>     ***`=` no longer used as assignment operator***
>
> -   Consider avoid the use of sapply. See this
>     <https://adv-r.hadley.nz/functionals.html>. For instance, consider
>
> ```
> y <- y[, apply(!is.na(y), 2, sum) > 3, drop = FALSE]
> ```
>
> instead of
>
> ```
> y <-
>                y[, sapply(y, function(col)
>                     sum(!is.na(col))) > 3, drop = FALSE]
> ```
>
> ***`sapply()` no longer used in this, or other functions***
>
> #### `sw_model.R`
>
> -   Ensure uniformity in assignment by using the same operator
>     consistently---avoid mixing `<-` and `=`. eg.
>     `n_obs = sum(observed$count)`
>
>     ***`=` no longer used as assignment operator***
>
> -   I'm not sure if incorporating the reading of a csv is optimal in a
>     function. What if a user wants to input a dataframe that isn't
>     from a CSV file? I propose modifying the function to accept a
>     dataframe or tibble with mandatory `n_sapwood` and `count`
>     columns. Additionally, allowing users to import from a CSV could
>     be achieved by either creating a small auxiliary function or
>     simply utilizing utils::read.csv(). This also reduce the size of
>     the sw_model function. Please consider this for all those
>     functions with the same issue.
>
>     ***You are right! I might have over complicated the sw-functions
[…]
