### 2024-03-20 #618 @hanecakr
It was suggested to rename the package to `fellingdater` instead of `fellingdateR`. Any tips on how to do this without breaking links and initializing git issues? @maelle @ajpelu @njtierney
