### 2024-03-20 #618 @hanecakr
I have a first draft of a paper ["paper/paper.md"] that I want to submit to JOSS if/when I manage to address all issues raised in the review proces. Can this paper remain a .md file under the `paper`-folder, should I remove it from the package until after review by JOSS, or should I already reformat it to a vignette?
