### 2024-03-21 #618 @maelle
@hanecakr Regarding renaming, you're in luck, @njtierney wrote a post on this very topic years ago: https://www.njtierney.com/post/2017/10/27/change-pkg-name/ (linked from the [dev guide](https://devguide.ropensci.org/pkg_building.html#naming-your-package))
