+++ tests/testthat.R
@@ -7,6 +7,6 @@
 # * https://testthat.r-lib.org/reference/test_package.html#special-files
 
 library(testthat)
-library(fellingdateR)
+library(fellingdater)
 
-test_check("fellingdateR")
+test_check("fellingdater")
