+++ pkgdown.yml
@@ -3,8 +3,8 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   getting_started: getting_started.html
-last_built: 2024-03-20T15:13Z
+last_built: 2024-03-21T10:23Z
 urls:
-  reference: https://hanecakr.github.io/fellingdateR/reference
-  article: https://hanecakr.github.io/fellingdateR/articles
+  reference: https://hanecakr.github.io/fellingdater/reference
+  article: https://hanecakr.github.io/fellingdater/articles
 
