+++ NEWS.md
@@ -1,3 +1,5 @@
+# fellingdater (development version)
+
 
 fellingdater 0.0.0.9003 (2023-11-20)
 ====================================
@@ -8,10 +10,12 @@ fellingdater 0.0.0.9003 (2023-11-20)
   * First draft op paper for Journal of Open Software in /paper/paper.md
   
   
-fellingdater 0.0.0.9004 (2024-03-20)
+fellingdater 0.0.0.9004 (2024-03-21)
 ====================================
+
 ### REVIEW
   * new version of the package after review
   * review history: https://github.com/ropensci/software-review/issues/618
+  * `fellingdateR`has been renamed `fellingdater`
   
  
