### 2024-03-21 #618 @maelle
@hanecakr thank you! I'd have recommended the opposite (quotes using the quote syntax, your comments as is) but whatever works!

@ajpelu thank you! Could you please post your answer with the [reviewer's approval template](https://devguide.ropensci.org/approval2template.html)?

@njtierney could you please have a look at the answer to your review?

Thanks all for all your work :pray:
