### 2024-03-21 #618 @ajpelu
## Reviewer Response
@hanecakr Thanks for your message. I have seen your changes, and for me it is fine. If you need help to improve the pkg, please let me know. We may want to collaborate on future versions of the package, as this topic (i.e. combining dendrochronology and R) is in line with my research interests. Also if you need any help with the manuscript, let me know.

@maelle Thanks for the opportunity to review this pkg. It was an inspiring experience.

#### Final approval (post-review)

Estimated hours spent reviewing: 13 h
