+++ inst/paper/paper.bib
@@ -467,6 +467,17 @@
 	langid = {en}
 }
 
+@article{bronkramsey1995,
+  title = {Radiocarbon Calibration and Analysis of Stratigraphy: The {{OxCal}} Program},
+  author = {Bronk Ramsey, C.},
+  year = {1995},
+  journal = {Radiocarbon},
+  volume = {37},
+  number = {2},
+  pages = {425--430},
+  doi = {10.1017/S0033822200030903}
+}
+
 @article{brewer2011,
 	title = {Summary of dendro data formats (published as supplementary material for Brewer, Murphy & Jansma, 2011. TRiCYCLE: a universal conversion tool for digital tree-ring data).},
 	author = {Brewer, P. and Murphy, D.},
@@ -481,7 +492,7 @@
 
 @article{crema2020,
 	title = {Inference from large sets of radiocarbon dates: software and methods},
-	author = {Crema, Enrico R and Bevan, Andrew},
+	author = {Crema, E.R. and Bevan, A.},
 	year = {2020},
 	month = {10},
 	date = {2020-10-06},
@@ -490,4 +501,28 @@
 	doi = {10.1017/RDC.2020.95},
 	url = {https://www.cambridge.org/core/product/identifier/S0033822220000958/type/journal_article},
 	langid = {en}
+}
+
+@article{dominguez-delmas_2020,
+	title = {Seeing the forest for the trees: {New} approaches and challenges for dendroarchaeology in the 21st century},
+	volume = {62},
+	doi = {10.1016/j.dendro.2020.125731},
+	language = {en},
+	urldate = {2020-09-08},
+	journal = {Dendrochronologia},
+	author = {Domínguez-Delmás, M.},
+	year = {2020},
+	pages = {125731},
+     langid = {en}
+}
+
+@article{shindo2024,
+  title = {European Larch Sapwood: {{A}} Model for Predicting the Cambial Age and for a More Accurate Dating},
+  author = {Shindo, L. and Saulnier, M. and Raese, H. and Guibal, F. and Edouard, J.-L. and Bolka, M. and Carrer, M. and Corona, C. and Gassmann, P. and Grabner, M. and Guillet, S. and Nicolussi, K. and Nola, P. and Pignatelli, O. and Stoffel, M.},
+  year = {2024},
+  journal = {Dendrochronologia},
+  volume = {83},
+  pages = {126150},
+  doi = {10.1016/j.dendro.2023.126150},
+  langid = {en}
 }
