+++ inst/paper/paper.md
@@ -1,10 +1,10 @@
 ---
-title: "_fellingdater_: an R package to estimate, report and combine felling dates of historical tree-ring series."
+title: "_fellingdater_: a toolkit to estimate, report and combine felling dates derived from historical tree-ring series."
 author: Kristof Haneca
 orcid: 0000-0002-7719-8305
 affiliation: Flanders Heritage Agency, Brussels, Belgium
 email: kristof.haneca@vlaanderen.be
-date: 2023-11-17
+date: 2024-04-21
 output: 
   pdf_document: default
   html_document:
@@ -24,13 +24,13 @@ or a piece of timber. It involves measuring the width of each growth
 ring and comparing the measured ring-width pattern to absolutely dated
 reference chronologies. Once a tree-ring series is securely anchored to
 a calendar year time-scale, the end date of the outermost ring can be
-used to determine or estimate the year of death of the parent tree (i.e.
-the felling of the tree).
+used to determine or estimate the year of death of the parent tree
+(i.e., the felling of the tree).
 
-The `fellingdater` package aims to offer a set off functions that can
+The `fellingdater` package aims to offer a suite of functions that can
 assist dendrochronologists to infer, combine and report felling date
 estimates from dated tree-ring series of (pre-)historical timbers, based
-on the presence of (partially) preserved sapwood or waney edge ([Fig.
+on the presence of partially preserved sapwood or waney edge ([Fig.
 1](#fig-cross-section)).
 
 ![A cross-section of a historical timber from a medieval roof
@@ -52,43 +52,46 @@ be anchored to an absolute time scale. From a dated tree-ring pattern it
 is known in which growing season each growth ring has been laid down by
 the parent tree. In archaeological, architectural of art-historical
 studies, the primary objective of a dendrochronological survey is to
-establish an accurate estimate of the **felling date** (or dying-off) of
+ascertain an accurate estimate of the **felling date** (or dying-off) of
 the parent tree from which the timber originates [@haneca2009;
 @tegel2022]. This felling date is the closest related and datable event
-to the creation of the wooden object or construction under study.
+to the creation of the wooden object or construction under study. These
+precisely dated events serves as the basis for narratives on various
+aspects, such as timber selection, craftsmanship, workshop practices,
+trade, provenance, and historical forest management
+[@dominguez-delmas_2020].
 
 The exact felling date can be inferred from the calendar year assigned
 to the most recently formed tree ring prior to the felling or death of
-the tree. If this last-formed ring is present on the object or timber
-under study, tree-ring dating can reach (sub-)annual chronological
-resolution. Unfortunately, this prerequisite is often not fulfilled. The
-wood of the felled tree has been processed, trimmed, or has gone through
-a process of biological deterioration leading to the irreversible loss
-of wood tissue. When the outermost portion of the timber no longer
-includes the cambial zone (such as in [Fig. 1](#fig-cross-section)), the
-timing of the felling date can only be estimated. The most unfavorable
-situation is when no sapwood, and not even the transition between
-heartwood and sapwood, remains on the object or timber under study
-([Fig. 1](#fig-cross-section), HW/SW boundary). **Sapwood** includes the
-outermost wood tissues of the xylem of a living tree, and is the
-physiologically active outer portion of the stem or a branch. It is
-situated between the cambial zone and the (dead) heartwood, and includes
-several growth rings. If none of the sapwood is retained, an untraceable
-amount of wood and growth layers has been removed. The last measured and
-dated ring then only provides an earliest possible felling date or
-*terminus post quem*.
+the tree. Achieving this requires the presence of the last-formed ring
[…]
