+++ .github/workflows/draft-pdf.yml
@@ -20,4 +20,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: .inst/paper/paper.pdf
+          path: ./inst/paper/paper.pdf
