+++ inst/paper/paper.md
@@ -1,15 +1,23 @@
 ---
 title: "_fellingdater_: a toolkit to estimate, report and combine felling dates derived from historical tree-ring series."
-author: Kristof Haneca
-orcid: 0000-0002-7719-8305
-affiliation: Flanders Heritage Agency, Brussels, Belgium
-email: kristof.haneca@vlaanderen.be
+authors:
+  - name: Kristof Haneca
+    orcid: 0000-0002-7719-8305
+    affiliation: 1
+affiliations:
+  - name: Flanders Heritage Agency, Belgium
+    index: 1
 date: 2024-04-21
 output: 
   pdf_document: default
   html_document:
   df_print: paged
-tags: ["R", "tree-ring analysis", "dendrochronology", "sapwood"]
+tags: 
+  - R
+  - tree-ring analysis
+  - dendrochronology
+  - sapwood
+  - dendroarchaeology
 editor_options: 
   markdown: 
     wrap: 72
