### 2024-03-28 #618 @maelle
@njtierney could you please have a look at the answer to your review? Please use the [reviewer's approval template](https://devguide.ropensci.org/approval2template.html).
