### 2024-04-08 #618 @njtierney
## Reviewer Response

Thanks for the responses, @hanecakr! You have done a great job addressing these comments. Thank you so much for taking the time to implement these. Well done! I've got a couple of minor comments below but I think this is all good to go from my perspective.

Formatting these reviews can be a bit tricky, but I did find it hard to know which parts you were responding to - in particular the documentation parts. This could have been resolved by placing my entire review in block quotes and then you responding either in further nested block quotes (`> >`) or by responding outside the blockquote.

### README

The README looks to be updated substantially, thank you for taking the time to do this! One minor quirk I noticed with the README is:

> The presence of partially preserved sapwood (_sw_) allows to estimate the missing number of sapwood rings ( ? in figure below), and to report an interval in which the actual felling date (_fd_) likely falls.

I'm not sure what to make of the `? in figure below` - is that a typo?

I also really appreciate the new workflow diagram, that is very nice!

### package rename

Well done on renaming the package! One small quip - you still have the RStudio project named `fellingDateR` - I think this should be `fellingdater`.

### Using `<-`

There is one case in a test: `test-sw_model.R` where = is used.

#### Final approval (post-review)

Estimated hours spent reviewing: 6.5
