### 2024-04-08 #618 @ropensci-review-bot
Approved! Thanks @hanecakr for submitting and @njtierney, @ajpelu for your reviews! :grin:
