### 2024-04-09 #618 @hanecakr
Dear @maelle , @ajpelu and @njtierney thank you so much for all your time and effort you´ve invested in reviewing my package 🙏🙏🙏. It was a first for me and learned a lot. Thanks to you the quality of the package improved significantly!
I´m on a short family trip right now (without laptop 😅), but will address the remaining issues and transfer to rOpenSci´s repo next week.
My best wishes,
Kristof
