### 2024-04-16 #618 @hanecakr
@maelle, the transfer to rOpenSci's repo succeeded! 🚀
Shall I add you as a reviewer as well in DESCRIPTION?
Thanks again for all your time and guidance.
